% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxPositionWeights.R
\name{approxPositionWeights}
\alias{approxPositionWeights}
\title{Approximate Positional Bias}
\usage{
approxPositionWeights(se, assay_name = "fore", kmers, nk = 100L,
  smooth = TRUE, withse = TRUE, verbose = FALSE, log_scale = FALSE,
  offset = 1L, .smooth.span = 1/2, .filter = 1L, .trim = if
  (.filter > 0L) {     c(1, 36) } else {     NULL })
}
\arguments{
\item{se}{SummarizedExperiment of probe intensities}

\item{assay_name}{string name of the assay to use. (default = "fore")}

\item{kmers}{character vector of k-mers to predict.}

\item{nk}{number of top k-mers to use to estimate positional
trend for each sample. This should be larger for longer
k-mers and smaller for shorter k-mers as the number of probes
per k-mer depends strongly on k. (default = 100L)}

\item{smooth}{logical whether to return loess smoothed bias esimates along
with median bias estimates. If \code{withse} is also TRUE, each
position is inversely weighted by the estimated standard error.
If TRUE, will return list of tables, "median" and "smooth".
Loess smoothing can be adjusted using the \code{.smooth.span}
parameter. (default = FALSE)}

\item{withse}{logical whether to return standard error estimates along
with median trend estimates. If TRUE, will return list of tables,
"median" and "se". The standard error is estimated by the scaled MAD
divided by the square root of the number of observations at each
position. (default = FALSE)}

\item{verbose}{logical whether to print extra messages during model fitting
procedure. (default = FALSE)}

\item{log_scale}{logical whether to calculate bias on log2 scale
intensities. (default = FALSE)}

\item{offset}{integer offset to add to intensities before performing
log2 scale calculations. (default = 1L)}

\item{.smooth.span}{numeric span value to be passed to \code{stats::smooth}
to adjust amount of smoothing. The default value is set to a smaller
value than the default \code{stats::smooth} value. (default = 1/2)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating weights. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{.trim}{interger vector of length two specifying start and end
of probe sequence to be used. Default is based on the universal
PBM probe design where only leading 36nt should be used.
Ignored if NULL. (default = c(1, 36))}
}
\value{
A vector of weights of length equal to the M - K + 1, where M is the length
of the trimmed probe sequence, and K is the length of the K-mer sequences.
}
\description{
Probe intensities in protein binding microarrays (PBMs) are
highly dependent not only on the sequence content of each
probe, but also the position of the contained sequences within
the probe. Typically, if a motif is closer to the glass slide,
the binding ability of the TF will be reduced, and the resulting
probe intensity will be lower than if the same motif were
positioned within the probe closer to the free end of the
sequence. This function provides a simple way to estimate
the positional bias of a TF based on the relative probe intensities
for probes containing the top N K-mers, ranked by median K-mer
affinity.
}
\author{
Patrick Kimes
}
