% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay2tidy.R
\name{assay2tidy}
\alias{assay2tidy}
\title{SummarizedExperiment Assay to Tidy Object}
\usage{
assay2tidy(se, assay_name = "fore", long = FALSE, .filter = 1L)
}
\arguments{
\item{se}{SummarizedExpierment object.}

\item{assay_name}{string name of the assay to use. (default = "fore")}

\item{long}{logical whether to transform data to long format and
include colData in output rather than default wide format with
dimension similar to original SummarizedExperiment object.
(default = FALSE)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to plotting. (default = 1)}
}
\value{
tibble with assay data.
}
\description{
Simple helper function to convert assay format data in SummarizedExperiment
object into a tidy tibble/data.frame for interactive analysis.
}
\author{
Patrick Kimes
}
