% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundSubtract.R
\name{backgroundSubtract}
\alias{backgroundSubtract}
\title{Background subtract intensities}
\usage{
backgroundSubtract(pe, assay = SummarizedExperiment::assayNames(pe)[1],
  assayb = SummarizedExperiment::assayNames(pe)[2], keepb = TRUE,
  nonnegative = TRUE, verbose = FALSE)
}
\arguments{
\item{pe}{a PBMExperiment object containing PBM intensity data.}

\item{assay}{a numeric index or string specifying the foreground assay.
(default = \code{SummarizedExperiment::assayNames(pe)[1]})}

\item{assayb}{a numeric index or string specifying the background assay.
(default = \code{SummarizedExperiment::assayNames(pe)[2]})}

\item{keepb}{a logical value whether to keep the background assay
after subtraction. (default = TRUE)}

\item{nonnegative}{a logical value whether to restrict intensities to
non-negative values by setting negative values to NA. (default = TRUE)}

\item{verbose}{a logical value whether to print verbose output
during analysis. (default = FALSE)}
}
\value{
PBMExperiment object with background subtracted intensities replacing the original
\code{assay} assay values.
}
\description{
Given a PBMExperiment with foreground and background intensities,
this function performs simple background subtraction and returns the same
PBMExperiment object with the background intensities subtracted from the
foreground intensities.
}
\author{
Patrick Kimes
}
