% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPBMTable.R
\name{buildPBMTable}
\alias{buildPBMTable}
\title{Build PBM Sample Table}
\usage{
buildPBMTable(gpr_dir, gpr_type = "Alexa488")
}
\arguments{
\item{gpr_dir}{path to directory containing GPR files.}

\item{gpr_type}{string value specifying type of GPR files to find; must
be one of: "Alexa488", "Cy3", "Masliner", "all". (default = "Alexa488")}
}
\value{
tibble with each row corresponding to a sample GPR file.
}
\description{
Helper function to create table of GPR sample files from specified directory path.
Parsing of file names is based on previously observed patterns and may not be exact.
Expected naming conventions of GPR files are described in the Details below.
}
\details{
For the three types of GPR files currently supported, files are assumed to be named
according to the following conventions:
\itemize{
\item \strong{Alexa488}: \code{{date}_{version}_{runID}_{condition1}_...} \cr \code{_{condition8}_Alexa488_lp{laser}pg{gain}_{index}_8.gpr}
\item \strong{Masliner}: \code{{date}_{version}_{runID}_{condition1}_...} \cr \code{_{condition8}_Alexa488_MaslinerOutput_{index}_8.gpr}
\item \strong{Cy3}: \code{{date}_{version}_{runID}_{miscellaneous}_Cy3_lp{laser}pg{gain}_{index}_8.gpr}
}
}
\author{
Patrick Kimes
}
