% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upbm-validators.R
\name{checkKmers}
\alias{checkKmers}
\title{Check K-mer Sequences}
\usage{
checkKmers(kmers, verb)
}
\arguments{
\item{kmers}{character vector of k-mer sequences or \code{NULL}.}

\item{verb}{logical whether to print extra messages.}
}
\value{
Original \code{se} object with default \code{pbm_8x60k_v1} probe design added
if probe 'Sequence' column is missing and the 8x60k design dimensions match
to original \code{se}.
}
\description{
Simple helper function to check if specified k-mer set is valid,
namely a vector of equal length character strings.
If \code{NULL}, a default set of 8-mers is returned. If invalid, an error is thrown.
}
\author{
Patrick Kimes
}
