% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictKmers.R
\name{checkProbeSequences}
\alias{checkProbeSequences}
\title{Check Probe Sequences}
\usage{
checkProbeSequences(se, verb)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data and
probe sequence information in rowData, or simply the DataFrame/data.frame
corresponding to the rowData.}

\item{verb}{logical whether to print extra messages.}
}
\value{
Original \code{se} object with default \code{pbm_8x60k_v1} probe design added
if probe 'Sequence' column is missing and the 8x60k design dimensions match
to original \code{se}.
}
\description{
Simple helper function to check if probe sequences in the rowData of
a SummarizedExperiment object are valid. The sequences should be in
a column named 'Sequence'.
}
\author{
Patrick Kimes
}
