% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb2PBMExperiment.R
\name{comb2PBMExperiment}
\alias{comb2PBMExperiment}
\title{Combinatorial Files to PBMExerpiment Object}
\usage{
comb2PBMExperiment(tab)
}
\arguments{
\item{tab}{data.frame with at least a single \code{combn} column
corresponding to PBM "combinatorial.txt" file paths.}
}
\value{
SummarizedExperiment object containing PBM data.
}
\description{
Helper function to read in a set of combinatorial files generated by
the Universal PBM Analysis Suite into a format similar to raw PBM
data read using \code{buildPBMExerpiment}.
}
\details{
If a \code{name} column is included int he input table, these are used as the
corresponding IDs for each sample/row. If IDs are not provided, the
file names are used. Sample IDs must be unique, and an error is thrown
if any samples have a common ID.
All other columns (aside from \code{combn} and \code{name}) in \code{tab} will be added to
the \code{colData} of the returned SummarizedExperiment object. Any sample
information, e.g. replicate or condition information can be included in
\code{tab}.
}
\author{
Patrick Kimes
}
