% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy3factors.R
\name{cy3factors}
\alias{cy3factors}
\title{Compute Cy3 Scaling Factors}
\usage{
cy3factors(se, assay_name = "fore", .filter = 1L, verbose = FALSE,
  .trim = if (.filter > 0L) {     c(1, 36) } else {     NULL })
}
\arguments{
\item{se}{SummarizedExpierment object containing PBM Cy3 intensity data.}

\item{assay_name}{string name of the assay to use. (default = "fore")}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{.trim}{interger vector of length two specifying start and end
of probe sequence to be used. Default is based on the universal
PBM probe design where only leading 36nt should be used. 
Ignored if NULL. (default = c(1, 36))}
}
\value{
original SummarizedExperiment object with additional assays corresponding
to residuals, expected intensities, and ratio of observed vs. expected
probe intensities. Cy3 models are stored in the metadata of the returned
object.
}
\description{
PBM arrays are scanned twice, once for Cy3-tagged dUTPs to quantify
dsDNA abundance at each probe, and again for the Alexa488-tagged 
protein.
}
\author{
Patrick Kimes
}
