% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy3FitEmpirical.R
\name{cy3FitEmpirical}
\alias{cy3FitEmpirical}
\title{Compute Cy3 scaling factors using emprical reference}
\usage{
cy3FitEmpirical(pe, refpe,
  assay = SummarizedExperiment::assayNames(pe)[1], useMean = TRUE,
  standardize = TRUE, threshold = 1/2, verbose = FALSE)
}
\arguments{
\item{pe}{a PBMExperiment object containing Cy3 intensity data.}

\item{refpe}{a PBMExperiment object containing Cy3 reference
intensities as a single assay named \code{"ref"}.}

\item{assay}{a numeric index or string name specifying the assay to use.
(default = \code{SummarizedExperiment::assayNames(pe)[1]})}

\item{useMean}{a logical whether to use the probe-level mean, rather than
the probe-level median as the Cy3 reference (default = TRUE)}

\item{standardize}{a logical value whether to standardize residuals using
MAD (median absolute deviation about the median) intensity
computed for PBM Cy3 reference data. (default = TRUE)}

\item{threshold}{a numeric threshold on absolute value of log2 ratio between
observed and expected Cy3 intensities. If \code{standardize} is
specified, the log2 ratios are first scaled before comparing against
the threshold. In this case, the threshold is also scaled by the
median of the probe-level MAD intensities. (default = 1/2)}

\item{verbose}{a logical value whether to print verbose output during
analysis. (default = FALSE)}
}
\value{
Original Cy3 PBMExperiment object with additional assays corresponding
to ratio of observed to expected probe intensities, and whether probes were
flagged as low-quality based on \code{abs(log2(ratio)) > threshold}.
}
\description{
PBM arrays are scanned twice, once for Cy3-tagged dUTPs to quantify
dsDNA abundance at each probe, and again for the Alexa488-tagged 
protein. The Cy3 scans can be used to first filter out probes which
appear to have poor dsDNA enrichment, and second, to scale Alexa488
intensities to account for differences in dsDNA abundance between
probes.

Since probe intensities in the Cy3 scans are independent of the
assayed Alexa488-tagged protein, Cy3 scans should be comparable across
experiments. Based on this observation, a global Cy3 reference can be
constructed by taking the empirical average over a large number of Cy3
scans. Filtering and scaling can then be performed based on deviations
from this reference. Given a PBMExperiment of Cy3 scans and an
empirical reference generated by \code{cy3GenerateRef}, this function
returns the residuals from the reference, and the corresponding
observed-to-expected ratios for each probe as new assays added to the
original Cy3 PBMExperiment object.

The returned PBMExperiment object can be passed to \code{cy3Normalize}
to filter low quality probes and/or normalize Alexa488 intensities by the
computed ratios.
}
\seealso{
cy3GenerateRef cy3Normalize
}
\author{
Patrick Kimes
}
