% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy3Normalize.R
\name{cy3Normalize}
\alias{cy3Normalize}
\title{Perform Cy3 normalization}
\usage{
cy3Normalize(pe, cy3pe, assay = SummarizedExperiment::assayNames(pe)[1],
  match_by = "id_idx", filter = TRUE, scale = TRUE,
  verbose = FALSE)
}
\arguments{
\item{pe}{a PBMExperiment object containing Alexa488 intensity data.}

\item{cy3pe}{a PBMExperiment object containing Cy3 deviation values from
using \code{cy3FitEmpirical} or \code{cy3FitModel}.}

\item{assay}{a numeric index or string name specifying the assay to use.
(default = \code{SummarizedExperiment::assayNames(pe)[1]})}

\item{match_by}{a string column name in colData of \code{pe} and \code{cy3pe}
to use for matching scans across the two objects; values of
column must be unique for each scan in each object. (default = \code{"id_idx"})}

\item{filter}{a logical value whether to filter "low quality" flagged probes by setting to NA.
(default = TRUE)}

\item{scale}{a logical value whether to perform scaling by ratio of observed vs. expected
Cy3 intensity at each probe. (default = TRUE)}

\item{verbose}{a logical value whether to print verbose output during
analysis. (default = FALSE)}
}
\value{
Original PBMExperiment object with assay containing Cy3 filtered and/or scaled intensities
(\code{"normalized"}). If an assay with the same name is already included in the object, it will
be overwritten.
}
\description{
Given a PBMExperiment containing Alexa488 probe intensity data and a second
PBMExperiment containing Cy3 observed-to-expected ratios, this function
performs probe filtering and/or scaling of the Alexa488 data according to the
over/under abundance of dsDNA at each probe as quantified from the Cy3 data.

Both the Alexa488 and Cy3 PBMExperiment objects must have a common column in the
colData (specified by \code{match_by=}) that can be used to match scans across the
two objects. While the values of the \code{match_by} column may be repeated across
Alexa488 scans (e.g. if an array was reused or scanned multiple times), the
values must be unique across the Cy3 PBMExperiment object.
}
\seealso{
\code{\link{cy3FitEmpirical}}
}
\author{
Patrick Kimes
}
