% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerFit-estimators.R
\name{dl2_estimator}
\alias{dl2_estimator}
\title{Two-Step DerSimonian aand Kacker Estimator}
\usage{
dl2_estimator(Y, vi, k)
}
\arguments{
\item{Y}{probe effect sizes}

\item{vi}{probe variances}

\item{k}{number of probes}
}
\value{
list of estimates:
\itemize{
\item betaFE: effect size with no cross-study variance
\item varFE: total variance with no cross-study variance
\item betaME: effect size with cross-study variance
\item varME: total variance with cross-study variance
\item tau2: cross-study variance 
}
}
\description{
This is an implementation of the DerSimonian and Kacker two-step estimator of
cross-study variance, originally proposed as an improvement over the one-step
DerSimonian and Laird estimator in the context of meta analysis.
}
\references{
\itemize{
\item DerSimonian, R., & Kacker, R. (2007). Random-effects model for meta-analysis of clinical trials: an update. Contemporary Clinical Trials, 28(2), 105-114.
}
}
\author{
Patrick Kimes
}
