% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerFit.R
\name{dl_estimator}
\alias{dl_estimator}
\title{DerSimonian and Laird Estimator}
\usage{
dl_estimator(Y, vi, k)
}
\arguments{
\item{Y}{probe effect sizes}

\item{vi}{probe variances}

\item{k}{number of probes}
}
\value{
list of estimates:
\itemize{
\item betaFE: effect size with no cross-study variance
\item varFE: total variance with no cross-study variance
\item betaME: effect size with cross-study variance
\item varME: total variance with cross-study variance
\item tau2: cross-study variance 
}
}
\description{
This is an implementation of the DerSimonian and Laird one step estimator of
cross-study variance, originally proposed in the context of meta analysis,
adapted from the \code{metafor} package. The function is used to estimate
the cross-probe variance for each k-mer probe set.
}
\references{
\itemize{
\item DerSimonian, R., & Laird, N. (1986). Meta-analysis in clinical trials. Controlled Clinical Trials, 7(3), 177-188.
\item Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1-48. URL: http://www.jstatsoft.org/v36/i03/
}
}
\author{
Patrick Kimes
}
