% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCy3Empirical.R
\name{fitCy3Empirical}
\alias{fitCy3Empirical}
\title{Compute Cy3 Scaling Factors using Emprical Reference}
\usage{
fitCy3Empirical(se, refse, assay_name = "fore", standardize = TRUE,
  threshold = 1L, .filter = 1L)
}
\arguments{
\item{se}{SummarizedExpirement object containing PBM Cy3 intensity data.}

\item{refse}{SummarizedExperiment object containing PBM Cy3 reference
intensities.}

\item{assay_name}{string name of the assay to use. (default = "fore")}

\item{standardize}{logical whether to standardize residuals using
MAD (median absolute deviation about the median) intensity
computed for PBM Cy3 reference data. (default = TRUE)}

\item{threshold}{numeric threshold on absolute value of log2 ratio between
observed and expected Cy3 intensities. If \code{standardize} is
specified, the log2 ratios are first scaled before comparing against
the threshold. In this case, the threshold is also scaled by the
median of the probe-level MAD intensities. (default = 1L)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1L)}
}
\value{
original SummarizedExperiment object with additional assays corresponding
to ratio of observed vs. expected probe intensities, and whether probes were
flagged as low-quality based on abs(log2(ratio)) > \code{threshold}. Cy3 models are stored
in the metadata of the returned object.
}
\description{
PBM arrays are scanned twice, once for Cy3-tagged dUTPs to quantify
dsDNA abundance at each probe, and again for the Alexa488-tagged 
protein. Given a reference
}
\author{
Patrick Kimes
}
