% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCy3Models.R
\name{fitCy3Models}
\alias{fitCy3Models}
\title{Compute Cy3 Scaling Factors}
\usage{
fitCy3Models(se, assay_name = "fore", refit = TRUE, .filter = 1L,
  .trim = if (.filter > 0L) {     c(1, 36) } else {     NULL })
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM Cy3 intensity data.}

\item{assay_name}{string name of the assay to use. (default = "fore")}

\item{refit}{logical whether to filter outliers and refit trinucleotide
linear regression model. (default = TRUE)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{.trim}{interger vector of length two specifying start and end
of probe sequence to be used. Default is based on the universal
PBM probe design where only leading 36nt should be used. 
Ignored if NULL. (default = c(1, 36))}
}
\value{
original SummarizedExperiment object with additional assays corresponding
to the ratio of observed vs. expected probe intensities, and whether probes were
flagged as low-quality based on log2(ratio) > 1 or < -1. Cy3 models are stored
in the metadata of the returned object.
}
\description{
PBM arrays are scanned twice, once for Cy3-tagged dUTPs to quantify
dsDNA abundance at each probe, and again for the Alexa488-tagged 
protein. The Cy3 scans can be used to first filter out probes which
appear to have poor dsDNA enrichment, and second, to scale Alexa488
intensities to account for differences in dsDNA abundance between
probes. Since probe intensities in the Cy3 scans are roughly
proportional to the number of adenines in the probe sequence, the
original Universal PBM Analysis Suite proposed estimating scaling factors
by looking at the residuals of a linear regression fit using all
tri-nucleotide sequences starting with an adenine. Given a
SummarizedExperiment of Cy3 scans, this function fit the tri-nucleotide
linear regression models and returns the expected Cy3 intensities,
the residuals from the fits, and the corresponding observed-expected
ratios for each probe as new assays added to the original Cy3
SummarizedExperiment object.
}
\details{
The returned SummarizedExperiment object can be passed to \code{cy3Normalize}
to filter low quality probes and/or normalize Alexa488 intensities by the
computed ratios.
}
\seealso{
cy3Normalize
}
\author{
Patrick Kimes
}
