% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitProbeDistribution.R
\name{fitProbeDistribution}
\alias{fitProbeDistribution}
\title{fit probe distribution}
\usage{
fitProbeDistribution(se, assay_name = "fore", model = c("NormGam",
  "NormExp"), method = "mle", .filter = 1L)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR intensity information. Notice usually the intensity should
should be raw intensity in non-log scale.}

\item{assay_name}{string name of the assay to normalize. (default = "fore")}

\item{model}{character string specifying model for fitting intensityies. 
Must be one of "NormGam", "NormExp". (default = "NormGam")}

\item{method}{method used to estimate the parameters for normal+exponential model. See more details from 
\code{limma::normexp.fit}. (default = "mle")}

\item{.filter}{integer specifying level of probe filtering to
perform prior to normalization. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}
}
\value{
A matrix with rows as condition and cols as parameter
}
\description{
This is a collection of several models to fit the probe intensities. It applys
normal+exponential convolutional model from \code{limma::normexp.fit} and 
normal+gamma convolutional model from \code{NormalGamma::normgam.fit}. This function returns the
parameter estimate of corresponding mode.
}
\author{
Dongyuan Song, Patrick Kimes
}
