% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCy3Ref.R
\name{generateCy3Ref}
\alias{generateCy3Ref}
\title{Generate Cy3 Empirical Reference}
\usage{
generateCy3Ref(cy3se, assay_name = "fore", offset = 1L, register = TRUE,
  .filter = 1L)
}
\arguments{
\item{cy3se}{SummarizedExperiment object containing Cy3 scan data.}

\item{assay_name}{string name of assay to use. (default = "fore")}

\item{offset}{numeric offset to add to intensities before log2 transforming
to prevent errors with zero intensities. (default = 1L)}

\item{register}{logical whether to scale intensities across samples.
(default = TRUE)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}
}
\value{
SummarizedExperiment object with Cy3 probe-level reference metrics.
}
\description{
Cy3 normalization is performed using an empirically-derived
reference distribution computed using a large collection of
samples. By default, samples are first scaled to have a common
median intensity, and the reference probe-level intensities are
computed on the log2 scale. The reference is calculated on the
log2 scale rather than the raw scale because the Cy3 scans will be
used for filtering and scaling, actions which require examining
fold changes and ratios rather than raw intensity differences.
}
\details{
Scaling is performed by multiplying the raw intensities of each
sample by a scaling factor such that the median intensity of each
sample is equal to the median sample-median intensity of the original
intensities. This multiplicative scaling is equivalent to an additive
shift on the log2 scale.
}
\author{
Patrick Kimes
}
