% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpr2PBMExperiment.R
\name{gpr2PBMExperiment}
\alias{gpr2PBMExperiment}
\title{Load raw PBM scan data as PBMExperiment}
\usage{
gpr2PBMExperiment(scans, useMedian = TRUE, filterFlags = TRUE,
  readBackground = TRUE, probes = NULL, ...)
}
\arguments{
\item{scans}{a data.frame of GPR scan information with at least a single `gpr'
column, and preferrably a second `type' column. See details for more
information on the content of these columns.}

\item{useMedian}{a logical value whether to use median fluorescent intensity
for each probe rather than mean fluorescent intensity.
(default = TRUE)}

\item{filterFlags}{a logical value whether to replace intensity values at probes
flagged manually or automatically as being low quality
 (`Bad': -100, `Absent': -75, `Not Found': -50) with NA. (default = TRUE)}

\item{readBackground}{a logical value whether to also read in probe background
intensities. (default = TRUE)}

\item{probes}{a \code{\link[=PBMDesign-class]{PBMDesign}} object containing
the probe design used for the data. (default = NULL)}

\item{...}{optionally, probe design information can be directly specified using
\code{design=}, \code{probeFilter=}, \code{probeTrim=} parameters. These
will be ignored if \code{probes=} is specified.}
}
\value{
\code{\link[=PBMExperiment-class]{PBMExperiment}} object with assays containing
intensities from the GPR files specified in the table of scans. The object
includes the following assays:
* \code{fore} - foreground probe intensities,
* \code{back} - background probe intensities (unless \code{readBackground = FALSE}).

Scan metadata included in the data.frame are stored as colData.
Probe design information, if specified, is included as rowData.
}
\description{
This function reads PBM scan data from GenePix Results (GPR) files and returns
a single \code{\link[=PBMExperiment-class]{PBMExperiment}} object containing
foreground and optionally background probe intensity data in assays. Each
GPR is stored as a column in the object, with metadata specified in the
input \code{scans} table stored in the colData of the object.

If probe design information is also specified as a \code{\link[=PBMDesign-class]{PBMDesign}}
object to the optional \code{probes=} argument, this information is also stored
in the rowData of the returned PBMExperiment.
}
\details{
The primary argument, \code{scans}, must be a data.frame of GPR scan metadata.
Each row of the data.frame should correspond to a single GPR scan. At a minimum
the data.frame must include a single \code{"gpr"} column specifying the paths
to the corresponding GPR files to be read in. 

Additionally, GPR files are parsed differently depending on the file type which must be
one of ``Alexa", ``Cy3", or ``Masliner". Both ``Alexa" and ``Cy3" scans
are treated as untouched GPR files, while ``Masliner" files treated as Masliner-processed
GPR files. If possible the scan type of each file should be
specified in a \code{"type"} column of the \code{scans} data.frame.
If a \code{"type"} column is not included, all GPR files are assumed to be Alexa488
scans unless the file name in the \code{"gpr"} column contains ``Masliner".
}
\seealso{
\code{\link{PBMExperiment-class}}, \code{\link{PBMExperiment}}, \code{\link{PBMDesign-class}}
}
\author{
Patrick Kimes
}
