% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerEstimate.R
\name{kmerAggregate}
\alias{kmerAggregate}
\title{Aggregate K-mer Probe Sets}
\usage{
kmerAggregate(se, kmers, positionbias = TRUE, .filter = 1L,
  .trim = if (.filter > 0L) {     c(1, 36) } else {     NULL })
}
\arguments{
\item{se}{probe-level affinity summaries generated by \code{probeFit} with each
assay corresponding to a different group label.}

\item{kmers}{character vector of k-mers to predict.}

\item{positionbias}{logical whether to correct for bias due to position
of K-mer along probe sequence. (default = TRUE)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{.trim}{interger vector of length two specifying start and end
of probe sequence to be used. Default is based on the universal
PBM probe design where only leading 36nt should be used. 
Ignored if NULL. (default = c(1, 36))}
}
\value{
SummarizedExperiment of K-mer, probe pairs.
}
\description{
After performing probe-level summarization across samples, this function creates
a K-mer by probe mapping for a specified list of K-mers. By default, probe
affinities are corrected for position bias assuming a consistent position bias
pattern, conditional on mean probe intensity for each K-mer. The bias correction
is performed such that the mean probe intensity for each K-mer is unchanged.
The bias correction is performed to reduce the cross-probe variance within
each K-mer probe set for downstream analysis.
}
\author{
Patrick Kimes
}
