% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerEstimate.R
\name{kmerEstimate}
\alias{kmerEstimate}
\title{Estimate K-mers Probe Intensities}
\usage{
kmerEstimate(se, assay_name = NULL, kmers, .filter = 1L, .trim = if
  (.filter > 0L) {     c(1, 36) } else {     NULL })
}
\arguments{
\item{se}{probe-level affinity summaries geneated by \code{probeEstimate} with each
assay corresponding to a different group label.}

\item{assay_name}{string name of the assay to aggregate. If NULL, all assays are
aggregated. (default = NULL)}

\item{kmers}{character vector of k-mers to predict.}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{.trim}{interger vector of length two specifying start and end
of probe sequence to be used. Default is based on the universal
PBM probe design where only leading 36nt should be used. 
Ignored if NULL. (default = c(1, 36))}
}
\value{
SummarizedExperiment of estimated K-mer intensities.
}
\description{
After performing probe-level aggregation across samples, this function applies probe
set aggregation to obtain K-mer level estimates of affinity.
}
\author{
Patrick Kimes
}
