% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerFit.R
\name{kmerFit}
\alias{kmerFit}
\title{Fit k-mer Probe Set Models}
\usage{
kmerFit(se, kmers, positionbias = TRUE, method = c("dl", "dl2"),
  contrasts = TRUE, baseline = NULL, .filter = 1L, .trim = if
  (.filter > 0L) {     c(1, 36) } else {     NULL })
}
\arguments{
\item{se}{SummarizedExpeirment of probe-level affinity summaries generated by
\code{probeFit}.}

\item{kmers}{character vector of k-mers to predict.}

\item{positionbias}{logical whether to correct for bias due to position
of k-mer along probe sequence. (default = TRUE)}

\item{method}{character name of method to use for estimating cross-probe variance
in each k-mer probe set. Currently, the non-iterative DerSimonian-Laird ("dl")
and two-step Dersimonian-Laird ("dl2") methods are supported. (default = "dl")}

\item{contrasts}{logical whether to compute contrasts for all conditions against a
specified \code{baseline} column. (default = TRUE)}

\item{baseline}{character name of baseline condition to use to calculate contrasts
for all other conditions. Should match a column name of SummarizedExperiment object.
If NULL, the baseline condition is chosen by looking for a column ending in "ref" or "REF".
If a unique baseline column isn't found, an error is thrown. (default = NULL)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{.trim}{interger vector of length two specifying start and end
of probe sequence to be used. Default is based on the universal
PBM probe design where only leading 36nt should be used. 
Ignored if NULL. (default = c(1, 36))}
}
\value{
SummarizedExperiment of estimated k-mer affinities and differences.
}
\description{
After performing probe-level summarization across samples, this function applies probe
set aggregation to obtain k-mer level estimates of affinity for a specified list of
k-mers. By default, probe affinities are corrected for position bias. Bias
correction is performed such that the mean cross-probe intensity for each k-mer is
typically unchanged, and only serves to reduce the cross-probe variance within probe sets
for each k-mer.
}
\author{
Patrick Kimes
}
