% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerTestContrast.R
\name{kmerTestContrast}
\alias{kmerTestContrast}
\title{Test k-mer Contrasts}
\usage{
kmerTestContrast(se)
}
\arguments{
\item{se}{SummarizedExperiment of k-mer results from \code{kmerFit}.}
}
\value{
SummarizedExperiment of significance results for testing k-mer affinity
differences between conditions.
}
\description{
After estimating k-mer level affinities using probe-set aggregate, this
function tests for differential affinity between conditions.
}
\author{
Patrick Kimes
}
