% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerTestSpecificity.R
\name{kmerTestSpecificity}
\alias{kmerTestSpecificity}
\title{Test k-mer Specificities}
\usage{
kmerTestSpecificity(se, span = 0.05, ...)
}
\arguments{
\item{se}{SummarizedExperiment of k-mer results from \code{kmerFit}.}

\item{span}{span parameter to be passed to \code{limma::loessFit}.
(default = 0.05)}

\item{...}{other parameters to pass to \code{limma::loessFit}.}
}
\value{
SummarizedExperiment of significance results for testing k-mer specificity
differences between conditions.
}
\description{
After estimating k-mer level affinities using probe-set aggregate, this
function tests for differential specificity between conditions.
}
\author{
Patrick Kimes
}
