% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaBackgroundCorrect.R
\name{limmaBackgroundCorrect}
\alias{limmaBackgroundCorrect}
\title{SummarizedExperiment Wrapper for limma backgroundCorrect}
\usage{
limmaBackgroundCorrect(se, assay_name = "fore", assayb_name = NULL, ...,
  .force = FALSE, .nonnegative = TRUE)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data}

\item{assay_name}{string name of the assay to adjust. (default = "fore")}

\item{assayb_name}{string name of the background assay to use for
adjustment. (default = NULL)}

\item{...}{parameters to pass to \code{limma::backgroundCorrect.matrix}.
See details below for more information on main parameters.}

\item{.force}{logical whether to run correction even if data
has already been corrected for background. (default = FALSE)}

\item{.nonnegative}{logical whether to restrict intensities to non-negative
values by setting negative values to NA. (default = TRUE)}
}
\value{
SummarizedExperiment object with background corrected intensities.
}
\description{
Given PBM intensities stored as a SummarizedExperiment, this function
performs background correction on each sample using limma's
\code{backgroundCorrect} function, and returns the same
SummarizedExperiment object with the background corrected intensities.
}
\details{
The most relevant parameters specified to \code{limma::backgroundCorrect.matrix}
are \code{method =} and \code{normexp.method}.
Options for \code{method} include: "none" and "normexp". The default for
single-channel arrays is "normexp". Options for \code{norm.exp} include:
"saddle", "mle", "rma" and "rma75". The default is "saddle".
For other parameter, see the \code{limma::backgroundCorrect.matrix}
documentation.
}
\author{
Patrick Kimes
}
