% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaNormalizeBetweenArrays.R
\name{limmaNormalizeBetweenArrays}
\alias{limmaNormalizeBetweenArrays}
\title{SummarizedExperiment Wrapper for limma NormalizeBetweenArrays}
\usage{
limmaNormalizeBetweenArrays(se,
  assay = SummarizedExperiment::assayNames(se)[1], ..., .force = FALSE)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data}

\item{assay}{string name of the assay to adjust.
(default = \code{SummarizedExperiment::assayNames(se)[1]})}

\item{...}{parameters to pass to \code{limma::normalizeBetweenArrays}.
See details below for more information on main parameters.}

\item{.force}{logical whether to run normalization even if data
has already been normalized within arrays. (default = FALSE)}
}
\value{
SummarizedExperiment object with normalized intensities.
}
\description{
Given PBM intensities stored as a SummarizedExperiment, this function
performs normalization across samples using limma's
\code{normalizeBetweenArrays} function, and returns the same
SummarizedExperiment object with the normalized intensities.
}
\details{
The most relevant parameter specified to \code{limma::normalizeBetweenArrays}
is \code{method =}. Options include: "none", "scale", "quantile", and
"cyclic loess." The default for single-channel arrays is "quantile".
For other parameter, see the \code{limma::normalizeBetweenArrays}
documentation.
}
\author{
Patrick Kimes
}
