% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowertailNormalization.R
\name{lowertailNormalization}
\alias{lowertailNormalization}
\title{Lower-Tail Normalization (TESTING)}
\usage{
lowertailNormalization(se, assay_name = "fore", q = 0.4,
  stratify = condition, baseline = NULL, .fits = FALSE)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR
intensity information.}

\item{assay_name}{string name of the assay to plot.
(default = "fore")}

\item{q}{percentile between 0 and 1 specifying lower tail to use for
fitting truncated normal distribution. (default = 0.4)}

\item{stratify}{unquoted name of column in colData of SummarizedExperiment (or
'\code{sample}') to use for comparing samples; values in column must be
unique for each sample. (default = condition)}

\item{baseline}{string name of baseline condition to
compare other conditions against; if not specified, guessed by looking for
'ref' in any value of the stratifying variable. (default = NULL)}

\item{.fits}{logical whether to just return a table of fits rather
than the normalized SummarizedExperiment object. (default = FALSE)}
}
\value{
SummarizedExperiment object with normalized intensities in
new assay, \code{scaled}.
}
\description{
This is an experimental normalization procedure based on fitting
a truncated normal distribution to the lower tail of log2-scale probe
intensities to register intensities across experiments. The
intensities are shift/scale transformed on the log2-scale to match
a specified reference sample. \emph{This code is still under
testing and not a finalized normalization procedure}.
}
\author{
Patrick Kimes
}
