% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePBMExperiment.R
\name{makePBMExperiment}
\alias{makePBMExperiment}
\title{Read PBMExperiment Data}
\usage{
makePBMExperiment(tab, useMean = FALSE, useBackground = FALSE,
  probes = NULL)
}
\arguments{
\item{tab}{table of samples with at least `gpr` and `vers` columns
corresponding to GPR file paths and the corresponding PBM
design version}

\item{useMean}{logical whether to use mean fluorescent intensity
for each probe rather than median fluorescent intensity
(default = FALSE)}

\item{useBackground}{logical whether to use background subtracted
intensity rather than non-subtracted intensity
(default = FALSE)}

\item{probes}{data.frame containing probe sequences in a column, 'Sequence',
or a character vector specifying the probe sequences. If specified,
these values will be added to the rowData of the returned
SummarizedExperiment object (default = NULL)}
}
\value{
a SummarizedExperiment object with a single assay, 'gpr', containing the probe
intensities for the samples specified in the input 'tab'.
}
\description{
Read PBM data from a table containing paths to GPR files.
}
\author{
Patrick Kimes
}
