% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapKmers.R
\name{mapKmers}
\alias{mapKmers}
\title{Map K-mer Motifs to PBM Probe Sequences}
\usage{
mapKmers(pd, kmers)
}
\arguments{
\item{pd}{a PBMDesign object containing probe sequence information.}

\item{kmers}{character vector of K-mers (of equal length) to which
probes should be mapped.}
}
\value{
A table of k-mer sequences and probes, with each row
corresponding to an occurrence of a k-mer sequence on a probe.
The table inclues all columns of the input \code{pd} design
as well as columns indicating the orientation (\code{orient}) and
position (\code{pos}) of the k-mer in the probe.
}
\description{
This helper function takes a PBMDesign object and set of k-mer
sequences of interest, and returns a mapping between probes and 
k-mer sequences.
Since no probe filtering or trimming is performed as part of this
function, \code{pbmTrimProbes} and \code{pbmFilterProbes} should
be called before passing the PBMDesign object to this function.
}
\author{
Patrick Kimes
}
