% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeAcrossReplicates.R
\name{normalizeAcrossReplicates}
\alias{normalizeAcrossReplicates}
\title{Normalize across replicates}
\usage{
normalizeAcrossReplicates(pe,
  assay = SummarizedExperiment::assayNames(pe)[1], group = "id",
  stratify = "condition", baseline = NULL, verbose = FALSE, ...)
}
\arguments{
\item{pe}{SummarizedExperiment object containing GPR intensity information.}

\item{assay}{a string name of the assay to normalize.
(default = \code{SummarizedExperiment::assayNames(pe)[1]})}

\item{group}{a character string specifying a column in \code{colData(pe)} to use for grouping replicates.
(default = \code{"id"})}

\item{stratify}{a character string specifying a column in \code{colData(pe)} to use for determining
the unique baseline scan within each \code{group} and to match samples across values of \code{group}.
(default = \code{"condition"})}

\item{baseline}{a character string specifying the baseline condition in the \code{stratify} column to normalize
other conditions against within each \code{group}. If not specified and set to NULL, the baseline
value is guessed by looking for ``ref" in any value of the \code{stratify} column. If multiple
matching values are found, a warning is thrown and the first matching sample is used.
(default = NULL)}

\item{verbose}{a logical value whether to print verbose output during analysis. (default = FALSE)}

\item{...}{additional parameters to be passed to \code{qqslope} to compute
scaling factors.}
}
\value{
Original PBMExperiment object with assay containing cross-replicate normalized intensities
(\code{"normalized"}) and new columns added to the colData, \code{"acrossRepMultScale"} and \code{"acrossRepAddScale"},
containing the inverse of the log-scale multiplicative and additive scaling factors used to normalize intensities.
If an assay with the same name is already included in the object, it will be overwritten.
}
\description{
Universal PBM experiments are often performed with several conditions of interest,
e.g. various allelic variants of the same transcription factor, assayed on arrays of
the same plate, with few replicate plates (commonly 2 or 3). Within individual replicates (plates,
observed probe intensities can vary greatly across conditions for biologically
uninteresting reasons, such as concentration differences. To explicitly correct for these
differences, normalization is performed in two steps.

First, normalization is performed within replicates (plates). More detail on this procedure
can be found in the \code{\link{normalizeWithinReplicates}} documentation.

Second, normalization is performed across replicates (plates) with the assumption that
biologically uninteresting differences between replicates affect probe intensities
both multiplicatively and additively on the log-scale. A single log-scale multiplicative normalization
factor is first estimated for all samples within a replicate. Then, an log-scale additive normalization
is estimated such that the median intensities of the \code{baseline} samples in each replicate
are equal. More details on this calculation are provided below.
}
\details{
The following procedure is used to estimate the log-scale multiplicative factor
for each replicate. First, a cross-replicate reference is computed for each condition
(specified in the \code{stratify} column) by taking the cross-replicate mean quantiles 
of the observed log2 intensities. Next, a \emph{per-sample} log multiplicative scaling factor is
computed by taking the median ratio of the rank-ordered and median-centered log-probe intensities
between the sample and the corresponding reference distribution. Visually,
this can be interpreted as the approximate slope of the quantile-quantile (QQ) plot generated
using the log-scale intensities. The \emph{per-replicate} scaling factor is then computed
by taking the geometric mean of the per-sample factors across all samples
in the replicate (specified in the \code{group} column). To reduce the impact of outlier
probes, per-sample scaling factors are estimated using only the middle 80% of probe intensities.

After log-scale multiplicative factors have been estimated to correct for differences in
log-scale variance across replicates, a second log-scale additive factor is estimated
for each replicate to correct for differences in log-scale shift. A "global median" intensity
is first calculated across replicates by taking the geometric mean of the median
intensities in all \code{baseline} samples across replicates. This "global median" is computed
using the input probe intensities, i.e. without any cross-replicate normalization. 
The log-scale additive factor estimated as the difference between the median normalized probe
intensity of the \code{baseline} sample in each replicate and the "global median".
While the log-scale additive factor is estimated using only \code{baseline} samples, the normalization
is applied to all samples in the replicate.
}
\author{
Patrick Kimes
}
