% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeReplicates.R
\name{normalizeReplicates}
\alias{normalizeReplicates}
\title{Cross-Replicate Normalization}
\usage{
normalizeReplicates(se, assay = SummarizedExperiment::assayNames(se)[1],
  group = "id", stratify = "condition", baseline = NULL,
  pairwise = FALSE, onlyref = FALSE, ...)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR intensity information.}

\item{assay}{string name of the assay to normalize. (default = \code{SummarizedExperiment::assayNames(se)[1]})}

\item{group}{string name of column in colData of SummarizedExperiment to
use for grouping replicate experiments. (default = "id")}

\item{stratify}{string name of column in colData of SummarizedExperiment to
use for matching conditions across replicates; values in column must be
unique for each samples within each replicate group. (default = "condition")}

\item{baseline}{string name of baseline condition to
compare other conditions against; if not specified, guessed by looking for
'ref' in any value of the stratifying variable. (default = NULL)}

\item{pairwise}{logical whether scaling factors should be computed using all
pairwise comparisons or just using a single comparison against a median
quantile reference. See 'Details' for more information. (default = FALSE)}

\item{onlyref}{logical whether scaling factors should be computed using only
reference samples. (default = FALSE)}

\item{...}{additional parameters to be passed to \code{qqslope} to compute
scaling factors.}
}
\value{
SummarizedExperiment object with normalized intensities in
new assay, \code{repScaled}.
}
\description{
uPBM experimental replicates can often be on different scales, even
after log transformation. However, these differences are often consistent
across samples within replicates after \code{lowertailNormalization}.
Based on this observation, we perform cross-replicate normalization by
computing log-scale scaling factors at the level of replicates.
}
\details{
Scaling factors for each replicate is computed by comparing the
probe intensity quantiles across matched conditions in replicate
experiments. The replicate scaling factor is then computed as the
average scaling factor across all conditions. This procedure requires
performing all pairwise comparisons of matched conditions across all
pairs of replicate experiments. While rare in uPBM data, if the
number of replicate experiments is large (e.g. > 5), this procedure
can be slow. This is the default procedure (\code{pairwise = TRUE}).

Alternatively, the scaling factor can be computed using a single mean
quantile reference for each condition. This reference is taken as the
means of the quantile (and not the quantiles of the mean) probe intensities
for each condition across replicates. For each replicate, the scaling factor
is computed as the average scaling factor across all conditions compared
with the corresponding reference distribution. This can be faster when
the number of replicates is large, as it only requires a single comparison
for each sample in the data set. Specifying \code{pairwise = FALSE} will
use this approach.
}
\author{
Patrick Kimes
}
