% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeWithinReplicates.R
\name{normalizeWithinReplicates}
\alias{normalizeWithinReplicates}
\title{Normalize within replicates}
\usage{
normalizeWithinReplicates(pe,
  assay = SummarizedExperiment::assayNames(pe)[1], method = c("tmm",
  "quantile"), q = 0.6, qlower = 0, qdiff = 0.2, group = "id",
  stratify = "condition", baseline = NULL, verbose = FALSE)
}
\arguments{
\item{pe}{a SummarizedExperiment object containing GPR intensity information.}

\item{assay}{a string name of the assay to normalize.
(default = \code{SummarizedExperiment::assayNames(pe)[1]})}

\item{method}{a string specifying the method to use for normalization. Must be one of
\code{"tmm"} or \code{"quantile"}. Details on the methods are provided below.
(default = \code{"tmm"})}

\item{q}{a percentile between 0 and 1 specifying either the upper quantile of probes to include 
for normalization when \code{method = "tmm"} or the quantile to use for aligning samples
when \code{method = "quantile"}. (default = 0.6)}

\item{qlower}{a percentile between 0 and 1-\code{q} specifying the lower quantile of
probes to include for normalization when \code{method = "tmm"}. This parameter is
ignored when \code{method = "quantile"}. (defalut = 0)}

\item{qdiff}{a percentile between 0 and 0.5 specifying the additional fraction of lower-tail
probes to filter based on the deviation from the baseline condition when \code{method = "tmm"}.
Probes with the \code{qdiff} smallest (most negative) and the \code{qdiff} largest (most positive)
deviations from baseline condition will be filtered from normalization.
This parameter is ignored when \code{method = "quantile"}. (default = 0.2)}

\item{group}{a character string specifying a column in \code{colData(pe)} to use for grouping replicates.
If scans shouldn't be grouped, specify NULL. (default = \code{"id"})}

\item{stratify}{a character string specifying a column in \code{colData(pe)} to use for determining
the unique baseline scan within each \code{group}. (default = \code{"condition"})}

\item{baseline}{a character string specifying the baseline condition in the \code{stratify} column to normalize
other conditions against within each \code{group}. If not specified and set to NULL, the baseline
value is guessed by looking for ``ref" in any value of the \code{stratify} column. If multiple
matching values are found, one value is chosen arbitrarily. If the baseline condition is missing
from any \code{group} with more than one scan, an error is thrown. (default = NULL)}

\item{verbose}{a logical value whether to print verbose output during analysis. (default = FALSE)}
}
\value{
Original PBMExperiment object with assay containing within-replicate normalized intensities
(\code{"normalized"}) and a new column added to the colData, \code{"withinRepScale"},
containing the inverse of the scaling factors used to normalize intensities.
If an assay with the same name is already included in the object, it will be overwritten.
}
\description{
Universal PBM experiments are often performed with several conditions of interest,
e.g. allelic variants, assayed on separate arrays of the same plate with few replicates.
Within and across plates, probe intensities can vary for biologically
uninteresting reasons, such as concentration differences. To explicitly correct for these
differences, normalization is performed in two steps.

First, normalization is performed within replicates (plates) with the assumption that
biologically uninteresting differences only affect probe intensities multiplicatively.
Normalization factors are estimated for each sample relative to a baseline condition on each
plate. The baseline should ideally be a replicate wild type or other natural reference condition
included in each replicate (plate). This function includes approaches for performing this step of
normalization.

Second, normalization is performed across replicates (plates). More detail on this procedure 
can be found in the \code{\link{normalizeAcrossReplicates}} documentation.

The approaches to normalization implemented in this function make a fundamental assumption
that lower-tail probe intensities are distributed similarly across the conditions being
normalized. This assumption is generally satisfied for allelic variants of the same transcription
factor or transcription factors with similar binding affinities. However, this assumption may
not always hold, e.g. if comparing proteins of completely different families. \emph{In these cases,
normalization should be performed with caution, and analyses and plots comparing the distributions
of lower-tail probe intensities should be explored.}
}
\details{
The trimmed mean of M-values (\code{"tmm"}) method implemented in this function for cross-sample normalization
within replicates is based on the popular TMM method for RNA-seq data included
in the \code{edgeR} package. Very simply, a normalization factor is estimated as the trimmed mean
of probe-level log-scale differences between the baseline condition and sample using the lower
\code{[qlower, q]} percentile probes. Probes are ordered by the log-scale average intensity across
the baseline condition and sample. The trimmed mean is calculated excluding the top and bottom \code{qdiff}
probes.

Unlike RNA-seq expression estimates, PBM data show near-constant variance in log-scale differences
as a function of the log-scale mean intensities. Therefore, a simplified variant of the original
TMM method is used, where precision weights are not introduced.

The quantile-based (\code{"quantile"}) method \emph{should not be confused with what is commonly referred to
as ``quantile normalization."} Here, quantile-based normalization computes scaling factors across
}
\seealso{
\code{\link{normalizeAcrossReplicates}}
}
\author{
Dongyuan Song, Patrick Kimes
}
