% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pbm_8x60k_v1.R
\docType{data}
\name{pbm_designs}
\alias{pbm_designs}
\alias{pbm_8x60k_v1}
\alias{pbm_4x44k_v1}
\alias{pbm_4x44k_v2}
\alias{pbm_8x15k_v1}
\alias{pbm_8x15k_v2}
\title{Standard PBM Designs}
\format{data.frame with each row corresponding to a spot on a PBM array}
\source{
\url{http://the_brain.bwh.harvard.edu/PBMAnalysisSuite/}
}
\description{
Designs for standard PBM arrays are provided for convenience. For each
spot, the Row, Column, ID and Sequence information is provided.
}
\section{Designs}{

  pbm_8x60k_v1: Agilent array design 030265


  pbm_4x44k_v1: Agilent array design: 051681


  pbm_4x44k_v2: Agilent array design: 016060


  pbm_8x15k_v1: Agilent array design: 015803


  pbm_8x15k_v2: Agilent array design: 016236
}

\keyword{data}
