% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/PBM-accessors.R
\name{PBMDesign-replace}
\alias{PBMDesign-replace}
\alias{PBMDesign<-}
\alias{PBMDesign<-,PBMExperiment,PBMDesign-method}
\title{Set design in PBMExperiment object}
\usage{
PBMDesign(object) <- value

\S4method{PBMDesign}{PBMExperiment,PBMDesign}(object) <- value
}
\arguments{
\item{object}{a \code{\link[=PBMExperiment-class]{PBMExperiment}} object.}

\item{value}{a \code{\link[=PBMDesign-class]{PBMDesign}} object.}
}
\value{
modified PBMExperiment object
}
\description{
Add or replace \code{\link[=PBMDesign-class]{PBMDesign}} in
\code{\link[=PBMExperiment-class]{PBMExperiment}} object.
}
\seealso{
\code{\link{PBMDesign}}
}
\author{
Patrick Kimes
}
