% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{PBMDesign<-}
\alias{PBMDesign<-}
\title{Set design in PBMExperiment object}
\usage{
PBMDesign(x) <- value
}
\arguments{
\item{x}{\code{\link[=PBMExperiment-class]{PBMExperiment}} object.}

\item{value}{\code{\link[=PBMDesign-class]{PBMDesign}} or \code{NULL}.}
}
\value{
modified PBMExperiment object
}
\description{
Add, remove or place \code{\link[=PBMDesign-class]{PBMDesign}} in
\code{\link[=PBMExperiment-class]{PBMExperiment}} object.
Design can be removed by setting the value to \code{NULL}.
}
\seealso{
\code{\link{PBMDesign}}
}
\author{
Patrick Kimes
}
