% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PBMExperiment-class}
\alias{PBMExperiment-class}
\title{PBMExperiment class}
\description{
Extension of the \code{\link{SummarizedExperiment}} class to store
universal protein binding microarray (uPBM) data for use with functions
in the \pkg{upbm} package.
}
\author{
Patrick Kimes
}
