% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMExperiment-package.R
\docType{package}
\name{PBMExperiment-package}
\alias{PBMExperiment-package}
\alias{PBMExperiment}
\title{PBMExperiment}
\description{
This package implements the PBMExperiment class and associated methods for the
organization and analysis of transcription factor binding affinities quantified
in vitro using single-channel Protein Binding Microarrays.
}
\author{
Patrick Kimes
}
\keyword{package}
