% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMExperiment-class.R
\name{PBMExperiment}
\alias{PBMExperiment}
\title{Constructor function for PBMExperiment objects.}
\usage{
PBMExperiment(assays, colData = NULL, rowData = NULL, ...)
}
\arguments{
\item{assays}{a list containing.}

\item{colData}{a \code{\link{DataFrame}}.}

\item{rowData}{a \code{\link{DataFrame}}.}

\item{...}{additional parameters passed to \code{\link{SummarizedExperiment}}.}
}
\value{
a \code{\link{PBMExperiment}} object.
}
\description{
Function to construct \code{PBMExperiment} objects.
}
\author{
Patrick Kimes
}
