% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmFilterProbes.R
\name{pbmFilterProbes}
\alias{pbmFilterProbes}
\title{Filter PBMExperiment and PBMDesign objects}
\usage{
pbmFilterProbes(x)
}
\arguments{
\item{x}{a PBMExperiment or PBMDesign object}
}
\value{
Filtered object of same class as input \code{se}.
}
\description{
Not all probes in a PBM experiment should be used for analysis.
Given a PBMExperiment or a PBMDesign object with defined probe
filters, this function returns the subset of probes passing all
filtering criteria.
}
\author{
Patrick Kimes
}
