% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmPlotChip.R
\name{pbmPlotChip}
\alias{pbmPlotChip}
\title{Plot of Intensities on Grid}
\usage{
pbmPlotChip(se, assay = SummarizedExperiment::assayNames(se)[1],
  log_scale = TRUE, relative_scale = NULL, bound = Inf,
  .facet = TRUE)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR
intensity information.}

\item{assay}{string name of the assay to plot.
(default = \code{SummarizedExperiment::assayNames(se)[1]})}

\item{log_scale}{logical whether to plot the intensities
on the log-scale. (default = TRUE)}

\item{relative_scale}{string name of column to in \code{colData}
or numeric vector to use for scaling intensities in each
sample. This can be useful for plotting spatial bias of
samples relative to the total intensity in the sample. 
Ignored if NULL. (default = NULL)}

\item{bound}{non-negative numeric value to use for bounding outliers.
When specified, values are bound to \code{(-bound, bound)}.
If set to 0L, results are binarized to positive, negative.
Ignored if \code{Inf}. (default = Inf)}

\item{.facet}{logical whether plot should be faceted using
the default 'condition' column from the colData of the
SummarizedExperiment. (default = TRUE)}
}
\value{
ggplot object
}
\description{
Probe-level intensities plotted by row and column indices
defined in the \code{rowData} of the SummarizedExperiment
object. This can be useful for visually inspecting spatial
trends and biases across the scanned chip.
}
\details{
If the values in the specified assay include negative
values, \code{log_scale = FALSE} will be used. Note that
this behavior is different from the other plotting functions
in this package. To get around this, manually set all negative
values to NA prior to plotting.
}
\author{
Patrick Kimes
}
