% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmPlotComparison.R
\name{pbmPlotComparison}
\alias{pbmPlotComparison}
\title{Paired Scatterplot of Intensities}
\usage{
pbmPlotComparison(se1, se2, assay_name = "gpr", match_by = condition,
  log_scale = TRUE, maplot = FALSE, .method = "auto", .filter = 0)
}
\arguments{
\item{se1}{first SummarizedExperiment object containing GPR
intensity information.}

\item{se2}{second SummarizedExperiment object containing GPR
intensity information to be compared with first experiment.}

\item{assay_name}{string name of the assay to plot.
(default = "gpr")}

\item{log_scale}{logical whether to plot the intensities
on the log-scale. (default = TRUE)}

\item{.method}{value passed to \code{method = } parameter of the
\code{ggplot2::geom_smooth} function for adding a smoothed
fit to each scatter plot; to prevent any line, set NULL.
(default = "auto")}

\item{.filter}{integer specifying level of probe filtering to
perform prior to plotting. (default = 0)}

\item{baseline}{string name of baseline condition to
compare other conditions against; ignored if \code{se2}
is also provided. (default = "Ref")}

\item{ma}{logical whether to plot MA plot rather than standard
scatter plot. (default = FALSE)}
}
\value{
ggplot object
}
\description{
Probe-level intensities plotted pairwise between two replicates which
are matched using the \code{match_by} column of the colData. Only pairs of
samples which are present in both experiments are plotted.
}
\details{
As with \code{pbmPlotScatter}, if \code{maplot} is TRUE, a MA plot is created
with the same data, plotting A (log-scale mean intensity) against M (log-scale
difference in intensities).


Currently, function only includes capability of plotting
pairwise comparison of two experiment sets. This was initially
implemented with the understanding that rarely more than
two replicates of a set of experiments are performed with
the PBM technology. If the need becomes clear, this function
will be extended to accept an arbitrary number of
SummarizedExperiment objects.
}
\author{
Patrick Kimes
}
