% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmPlotDensity.R
\name{pbmPlotDensity}
\alias{pbmPlotDensity}
\title{Plot of Intensity Densities}
\usage{
pbmPlotDensity(se, assay = SummarizedExperiment::assayNames(se)[1],
  log_scale = TRUE, .facet = TRUE)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR
intensity information.}

\item{assay}{string name of the assay to plot.
(default = \code{SummarizedExperiment::assayNames(se)[1]})}

\item{log_scale}{logical whether to plot the intensities
on the log-scale. (default = TRUE)}

\item{.facet}{logical whether plot should be faceted using
the default 'condition' column from the colData of the
SummarizedExperiment. (default = TRUE)}
}
\value{
ggplot object
}
\description{
Densities of all probe-level intensities are plotted.
}
\author{
Patrick Kimes
}
