% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmPlotScatter.R
\name{pbmPlotScatter}
\alias{pbmPlotScatter}
\title{Scatterplot of Intensities}
\usage{
pbmPlotScatter(se, assay_name = "gpr", stratify = condition,
  baseline = NULL, log_scale = TRUE, maplot = FALSE, .method = "auto",
  .filter = 0)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR
intensity information.}

\item{assay_name}{string name of the assay to plot.
(default = "gpr")}

\item{stratify}{unquoted name of column in colData of SummarizedExperiment (or
'\code{sample}') to use for comparing samples; values in column must be
unique for each sample. (default = condition)}

\item{baseline}{string name of baseline condition to
compare other conditions against; if not specified, guessed by looking for
'ref' in any value of the stratifying variable. (default = NULL)}

\item{log_scale}{logical whether to plot the intensities
on the log-scale. (default = TRUE)}

\item{maplot}{logical whether to plot MA plot rather than standard
scatter plot. (default = FALSE)}

\item{.method}{value passed to \code{method = } parameter of the
\code{ggplot2::geom_smooth} function for adding a smoothed
fit to each scatter plot; to prevent any line, set NULL.
(default = "auto")}

\item{.filter}{integer specifying level of probe filtering to
perform prior to plotting. (default = 0)}
}
\value{
ggplot object
}
\description{
Probe-level intensities plotted pairwise against a reference
condition. Given a baseline condition specified with \code{baseline=}
(which must map uniquely in the \code{stratify=} column of the colData),
pairwise probe intensities are plotted with the identity line (x = y).
}
\details{
Alternatively, if \code{maplot} is TRUE, a MA plot is created with the
same data, plotting A (log-scale mean intensity) against M (log-scale
difference in intensities).


MA plots are drawn on the log scale by definition.
}
\author{
Patrick Kimes
}
