% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upbm-validators.R
\name{pbmProcessProbes}
\alias{pbmProcessProbes}
\title{Simple wrapper function to process PBM probe sequences}
\usage{
pbmProcessProbes(se, verbose = FALSE, .filter = 1L, .trim = c(1, 36))
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data and
probe sequence information in rowData, or simply the DataFrame/data.frame
corresponding to the rowData.}

\item{verbose}{logical whether to print extra messages. (default = FALSE)}

\item{.filter}{integer specifying level of probe filtering to perform. See
\code{pbmFilterProbes} for more details about levels of probe filtering.
(default = 1L)}

\item{.trim}{interger vector of length two specifying start and end
of probe sequence \strong{to be used}. Default is based on the universal
PBM probe design where only leading 36nt should be used. 
Ignored if \code{NULL}. (default = \code{c(1, 36)})}
}
\value{
Original \code{se} object with only probes passing filter, with trimmed Sequence columns.
If the probe sequences are not valid, an error is thrown.
}
\description{
Simple wrapper function to process PBM probe sequences
}
\seealso{
trimProbeSequences pbmFilterProbes checkProbeSequences
}
\author{
Patrick Kimes
}
