% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pbmtab.R
\docType{data}
\name{pbmtab}
\alias{pbmtab}
\title{data.frame containing sample information for example PBM experiment}
\description{
data.frame containing sample information for example PBM experiment
}
\keyword{data}
