% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmTrimProbes.R
\name{pbmTrimProbes}
\alias{pbmTrimProbes}
\title{Trim probe sequences}
\usage{
pbmTrimProbes(pe)
}
\arguments{
\item{pe}{a PBMExperiment or PBMDesign object.}
}
\value{
Original PBMExperiment or PBMDesign object with trimmed sequences.
}
\description{
Not all nucleotides in the probe sequences should be used for analysis.
Given a PBMExperiment or a PBMDesign object, this function returns a
modified object with probe sequences trimmed according to the
associated \code{probeTrim} slot value.
}
\author{
Patrick Kimes
}
