% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictkmers.R
\name{predictkmers}
\alias{predictkmers}
\title{Estimation of K-Mer Intensities}
\usage{
predictkmers(se, kmers = NULL, stdArray = TRUE, ...)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data}

\item{kmers}{character vector of k-mers to predict}

\item{stdArray}{logical whether the array is a standard PBM array. If
TRUE, then two processing steps will be applied prior to estimating
k-mer intensities, namely: 1. subsetting on de Bruijn sequences as
determined by the ID column of the SummarizedExperiment rowData, and
2. trimming of sequences to only use the unique 36nt region.
(default = TRUE)}

\item{...}{parameters to be passed to glmnet call - will overwrite
default paramaters}
}
\value{
SummarizedExperiment object with predicted affinities.
}
\description{
This function uses stuff to do stuff.
}
\details{
Default signature for the glmnet call is:
* family = 'gaussian'
* alpha = 0.5
* intercept = FALSE
* lambda = rev(0:50)
* lower.limits = 0
* thresh = 1e-10
}
\author{
Patrick Kimes
}
