% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeFit.R
\name{probeFit}
\alias{probeFit}
\title{Fit Probe Models}
\usage{
probeFit(se, assay = SummarizedExperiment::assayNames(se)[1],
  groups = condition, offset = 1L, .filter = 1L, ...)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data.}

\item{assay}{string name of the assay to use. (default = \code{assayNames(se)[1]})}

\item{groups}{unquoted name of column in colData of SummarizedExperiment (or
'\code{sample}') for grouping samples; values in columns will be
used as group labels. If no group labels are available, all samples
can be pooled by specifying \code{NULL}. (default = condition)}

\item{offset}{integer offset to add to intensities before log2 scaling to
prevent errors with zero intensities. If set to 0, probes with
zero intensities are dropped/ignored in estimation. (default = 1)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{...}{named arguments to be passed to \code{limma::eBayes}. See
Details for default parameters used in this function (different from
\code{limma::eBayes} default).}
}
\value{
SummarizedExperiment object with probe-level intensity information aggregated
across replicates. Each assay contains probe-level information for a single
group in the specified \code{groups} variable.
}
\description{
This function uses limma to estimate probe-level mean and moderated
variance estimates across replicates within conditions.
}
\author{
Patrick Kimes
}
