% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsmoothNormalization.R
\name{qsmoothNormalization}
\alias{qsmoothNormalization}
\title{qsmooth Normalization (TESTING)}
\usage{
qsmoothNormalization(se_list, ref_rep = 1, assay_name = "fore",
  scaling_between_rep = TRUE, scaling_within_rep = TRUE, q = 0.5,
  .force = FALSE, .filter = 1L, plot_weight = TRUE, ...)
}
\arguments{
\item{se_list}{a list object of SummarizedExperiment objects containing GPR intensity 
information.}

\item{ref_rep}{integer specifying which replicate will be used for scaling before normalization. (default = 1)}

\item{assay_name}{string name of the assay to normalize. (default = "fore")}

\item{scaling_between_rep}{logical whether to scale two reps into same median. (default = TRUE)}

\item{scaling_within_rep}{logical whether to perform scaling within arrays in advance. (default = TRUE)}

\item{q}{percentile between 0 and 1 specifying the quantile to align different tfs within a rep.
See \code{lowertailNormalization}. (default = 0.5)}

\item{.force}{logical whether to run normalization even if data
has already been normalized within arrays. (default = FALSE)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to normalization. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{plot_weight}{logical whether to plot \code{qsmooth::qsmoothPlotWeights}. (default = TRUE)}

\item{...}{parameters to pass to \code{qsmooth::qsmooth}.
See details below for more information on main parameters.}
}
\value{
list of SummarizedExperiment object with normalized intensities in
new assay
}
\description{
Given a list of SummarizedExperiment objects containing GPR intensity 
information, this function performs normalization across samples using qsmooth's
\code{qsmooth} function, and returns the same list of
SummarizedExperiment objects with the normalized intensities.
Currently, since qsmooth does not allow inputs with missing values, this function just removes all
probes with NAs and shows a warning message. 
\emph{This code is still under testing}.
}
\author{
Dongyuan Song, Patrick Kimes
}
