% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMExperiment-class.R
\name{read_gpr}
\alias{read_gpr}
\title{Read Standard GPR File}
\usage{
read_gpr(f)
}
\arguments{
\item{f}{path to gpr file}
}
\value{
a tibble.
}
\description{
Simple wrapper for reading standard GPR files.
GPR files are tab-delimited with 35 header lines.
}
\author{
Patrick Kimes
}
\keyword{internal}
