% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePBMExperiment.R
\name{readGPR}
\alias{readGPR}
\title{Read GPR File as Assay}
\usage{
readGPR(x, useMean = FALSE, useBackground = FALSE)
}
\arguments{
\item{x}{path to GPR fle}

\item{useMean}{logical whether to use mean fluorescent intensity
for each probe rather than median fluorescent intensity
(default = FALSE)}

\item{useBackground}{logical whether to use background subtracted
intensity rather than non-subtracted intensity
(default = FALSE)}
}
\value{
SummarizedExperiment object with one assay with one column,
and corresponding rowData
}
\description{
Helper function for reading in each GPR file.
}
\details{
Since the name of the value column can vary across samples,
we pull it out using its index (5th column)
}
\author{
Patrick Kimes
}
