% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPBMExperiment.R
\name{readPBM}
\alias{readPBM}
\title{Read PBM Data File}
\usage{
readPBM(gpr_path, gpr_type, useMean = FALSE, filterFlags = TRUE,
  readBackground = TRUE)
}
\arguments{
\item{gpr_path}{path to GPR file.}

\item{gpr_type}{scan type; one of "Alexa", "Cy3", "Masliner", "RawData".}

\item{useMean}{logical whether to use mean fluorescent intensity
for each probe rather than median fluorescent intensity.
Ignored if \code{gpr_type = "RawData"}. (default = FALSE)}

\item{filterFlags}{logical whether to replace intensity values at probes
flagged manually or automatically as being low quality.
 ('Bad': -100, 'Absent': -75, 'Not Found': -50) with NA. (default = TRUE)}

\item{readBackground}{logical whether to also read in probe background
intensities. Ignored if \code{gpr_type = "RawData"}. (default = TRUE)}
}
\value{
See documentation for the appropriate function for more details.
}
\description{
Helper function for reading in a single PBM raw data file either
in GPR format or processed format, e.g. such as files uploaded
to the UniPROBE database.
If \code{gpr_type = "RawData"} is specified, \code{\link{readRawData}}
is used to read in the specified file. Otherwise, \code{\link{readGPR}}
is used. Parameters are passed directly to the appropriate function.
}
\seealso{
readGPR readRawData
}
\author{
Patrick Kimes
}
