% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMExperiment-constructor.R
\name{readRawData}
\alias{readRawData}
\title{Read Raw Data File}
\usage{
readRawData(gpr_path, filterFlags = TRUE, readBackground = TRUE)
}
\arguments{
\item{gpr_path}{path to text file containing GPR data. See above description
for more information on how this differs from \code{readGPR}.}

\item{filterFlags}{logical whether to replace intensity values at probes
flagged manually or automatically as being low quality.
('Bad': -100, 'Absent': -75, 'Not Found': -50) with NA.
(default = TRUE)}

\item{readBackground}{logical whether output should include a column of NAs
for background intensities. Only useful for matching column names with
raw GPR data. (default = TRUE)}
}
\value{
tibble (data.frame-like) object of a single raw data file with between three to six
columns: 'Column', 'Row', 'foreground', and if available, 'ID', 'Name', 'Sequence'.
Unlike with raw GPR files, the ID and Name columns are assumed to be correct in
raw data files.
}
\description{
Helper function for reading in a partially "cleaned" 'rawdata' file, e.g.
available on the UniPROBE database.
Raw GenePix scan output files typically include a 30 to 40 line header, as well
as a large number of columns with various probe-level metrics. However, data uploaded
to the UniPROBE database typically include a small subset of the columns, and no
header. This function is written to read in raw GPR data stored in the latter format.
Files in the first format can be read in using the \code{readGPR} function.
}
\details{
Columns included in \code{rawdata} files on UniPROBE can vary substantially.
Based on a scan of data sets available on the database, we assume that the
Alexa intensity columns correspond to median background subtracted foreground
intensities. The first column containing the string "flag" following the
Alexa intensity column is assumed to be the corresponding flags for the
Alexa scan.
}
\author{
Patrick Kimes
}
