% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmaBackgroundCorrect.R
\name{rmaBackgroundCorrect}
\alias{rmaBackgroundCorrect}
\title{RMA Background Correction of Samples}
\usage{
rmaBackgroundCorrect(se)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data}
}
\value{
SummarizedExperiment object with background corrected intensities.
}
\description{
Given PBM intensities stored as a SummarizedExperiment, this function
performs RMA-style background correction on each sample, and returns the
same SummarizedExperiment object with the background corrected intensities.
}
\author{
Patrick Kimes
}
