% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmaNormalize.R
\name{rmaNormalize}
\alias{rmaNormalize}
\title{RMA Normalization of Samples}
\usage{
rmaNormalize(se)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data}
}
\value{
SummarizedExperiment object with normalized intensities.
}
\description{
Given PBM intensities stored as a SummarizedExperiment, this function
performs RMA-style normalization on each sample (currently, only quantile
normalization is supported) and returns thesame SummarizedExperiment object
with the normalized intensities.
}
\author{
Patrick Kimes
}
