% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMDesign-show.R, R/PBMExperiment-show.R
\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{show,PBMDesign-method}
\alias{show,PBMExperiment-method}
\title{Show upbm objects}
\usage{
\S4method{show}{PBMDesign}(object)

\S4method{show}{PBMExperiment}(object)
}
\arguments{
\item{object}{a PBMDesign or PBMExperiment object to show.}
}
\description{
Show upbm objects
}
