% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-PBMExperiment.R
\name{tidy-PBMExperiment}
\alias{tidy-PBMExperiment}
\alias{tidy.PBMExperiment}
\title{Tidy PBMExperiment object}
\usage{
\method{tidy}{PBMExperiment}(x,
  assay = SummarizedExperiment::assayNames(x)[1], long = FALSE, ...)
}
\arguments{
\item{x}{a PBMExperiment object.}

\item{assay}{a numeric index or string name specifying the assay to use.
(default = \code{SummarizedExperiment::assayNames(x)[1]})}

\item{long}{a logical whether to transform data to long format and
include colData in output rather than default wide format with
dimension similar to original PBMExperiment object.
(default = FALSE)}

\item{...}{other parameters for \code{tidy} generic function.}
}
\value{
tibble with assay data.
}
\description{
Simple helper function to convert assay format data in PBMExperiment
object into a tidy tibble.
}
\author{
Patrick Kimes
}
