% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-SummarizedExperiment.R
\name{tidy-SummarizedExperiment}
\alias{tidy-SummarizedExperiment}
\alias{tidy.SummarizedExperiment}
\title{Tidy SummarizedExperiment Object}
\usage{
\method{tidy}{SummarizedExperiment}(x,
  assay = SummarizedExperiment::assayNames(x)[1], long = FALSE,
  .filter = 1L, ...)
}
\arguments{
\item{x}{SummarizedExperiment object.}

\item{assay}{string name of the assay to use.
(default = \code{assayNames(x)[1]})}

\item{long}{logical whether to transform data to long format and
include colData in output rather than default wide format with
dimension similar to original SummarizedExperiment object.
(default = FALSE)}

\item{.filter}{integer specifying level of probe filtering. (default = 1L)}

\item{...}{other parameters for \code{tidy} generic function.}
}
\value{
tibble with assay data.
}
\description{
Simple helper function to convert assay format data in SummarizedExperiment
object into a tidy tibble for interactive analysis.
}
\author{
Patrick Kimes
}
