% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerTest.R
\name{tidyProbeTest}
\alias{tidyProbeTest}
\title{Tidy Probe Intensity Differences Test Results}
\usage{
tidyProbeTest(fit, rowdat)
}
\arguments{
\item{fit}{limma result generated by \code{probeTest}}

\item{rowdat}{data.frame or DataFrame result generated by \code{probeTest}
containing metadata information for tests}
}
\value{
SummarizedExperiment object with probe-level testing results from model in
\code{fit}. Each assay contains the results for a single coefficient in the model.
}
\description{
This function tidys the results of probe-level tests returned by
\code{probeTest}. The original SummarizedExperiment object used to
perform the test must be specified.
}
\examples{
\dontrun{
res <- probeTest(myse, .fits = TRUE)
res <- tidyProbeTest(fit = res$fit, rowdat = res$rowdat)
}

}
\author{
Patrick Kimes
}
