% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmmNormalization.R
\name{tmmNormalization}
\alias{tmmNormalization}
\title{TMM Normalization}
\usage{
tmmNormalization(se, assay = SummarizedExperiment::assayNames(se)[1],
  q = 0.6, q0 = 0.2, qb = 0, stratify = "condition",
  baseline = NULL, .filter = TRUE, .fits = FALSE)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR intensity information.}

\item{assay}{string name of the assay to normalize.
(default = \code{SummarizedExperiment::assayNames(se)[1]})}

\item{q}{percentile between 0 and 1 specifying lower tail A-values to use for
normalization. (default = 0.6)}

\item{q0}{percentile between 0 and 0.5 specifying the fraction of M-values to be trimmed
from each side. If \code{q0 = 0.5} it yields median.  (default = 0.2)}

\item{qb}{percentile either between 0 and 1-\code{q} or equals \code{Auto}. If \code{qb}
is numeric then it specifies the proportion for lowest A-values that will
be removed; if \code{qb = "Auto"} then this cutoff will be the A-value of
the medians of two background intensities. (defalut = 0)}

\item{stratify}{string name of column in colData of SummarizedExperiment to
use for comparing samples; values in column must be
unique for each sample. Alternatively, can specify '\code{"sample"}' to
use column names. (default = "condition")}

\item{baseline}{string name of baseline condition to compare other conditions
against; if not specified, guessed by looking for
'ref' in any value of the stratifying variable. (default = NULL)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to normalization. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{.fits}{logical whether to just return a table of fits rather
than the normalized SummarizedExperiment object. (default = FALSE)}
}
\value{
SummarizedExperiment object with normalized intensities in
new assay, \code{scaled}.
}
\description{
This is the function using trimmed mean of M-values (TMM) method
to perform cross sample normalization. TMM normalization is very popular
for RNA-seq data normalization and was included
in package \code{edgeR}. However, here a modified TMM methods is used.
The difference is that weights are not introduced when calculating trimmed
mean since unlike RNA-seq, PBM data shows almost constant variance of M-value across A-value.
Also, for PBM data there is no library size.
}
\details{
For TMM normalization in \code{upbm}, the first step is trmming A-values
(one side by \code{q} or two sides by \code{q} and \code{qb}). The second step
is trimming M-values.
}
\author{
Dongyuan Song, Patrick Kimes
}
