% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upbm-validators.R
\name{trimProbeSequences}
\alias{trimProbeSequences}
\title{Trim probe sequences}
\usage{
trimProbeSequences(pe)
}
\arguments{
\item{pe}{a PBMExperiment object or PBMDesign object.}
}
\value{
Original PBMExperiment or PBMDesign object with trimmed sequences.
}
\description{
Simple helper function to trim probe sequences in the rowData of
a PBMExperiment or PBMDesign object.
}
\author{
Patrick Kimes
}
