% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueKmers.R
\name{uniqueKmers}
\alias{uniqueKmers}
\title{Generate Unique K-mer Sequences}
\usage{
uniqueKmers(k = 8L, .largek = FALSE)
}
\arguments{
\item{k}{an integer length of oligonucleotide sequences to be
returned. (default = 8L)}

\item{.largek}{a logical value whether to allow specifying a large \code{k},
i.e. \code{k > 10L}. (default = FALSE)}
}
\value{
Vector of unique k-mer strings.
}
\description{
This simple function returns a vector of all unique k-mers of
a specified length, filtered to remove reverse complement redundancy
such that, e.g. only one of "AA" and "TT" is kept. When deciding
between a k-mer and the reverse complement, the alphabetically earlier
sequence is kept. In the example above, "AA" will be kept because "AA"
occurs before "TT".

Note, this function is not meant to be used for \code{k > 10} as the
computational cost of the approach can quickly grow to be too much.
By default, an error willbe thrown if k > 10 is specified. This 
can be overriden by setting \code{largek = TRUE}.
}
\author{
Patrick Kimes
}
