% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upbm-package.R
\docType{package}
\name{upbm-package}
\alias{upbm-package}
\alias{upbm}
\title{upbm}
\description{
This package implements methods for the organization, pre-processing,
exploration, and analysis of transcription factor binding affinities quantified
in vitro using single-channel universal protein binding microarrays (uPBMs).
}
\author{
Patrick Kimes
}
\keyword{package}
