% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upbmPreprocess.R
\name{upbmPreprocess}
\alias{upbmPreprocess}
\title{PBM Preprocess}
\usage{
upbmPreprocess(
  pe,
  cy3pe,
  cy3refpe,
  assay = SummarizedExperiment::assayNames(pe)[1],
  params = list(),
  skip = vector("character", 0),
  verbose = 1L
)
}
\arguments{
\item{pe}{a PBMExperiment object containing Alexa488 intensity data.}

\item{cy3pe}{an optional PBMExperiment object containing Cy3 intensity data. If
corresponding Cy3 data is not available for the Alexa488 scans, this must be
set to NULL.}

\item{cy3refpe}{an optional PBMExperiment object containing Cy3 reference data to
use for Cy3 normalization with \code{\link{cy3FitEmpirical}} (recommended).
To perform model-based normalization with \code{\link{cy3FitModel}},
set to NULL. This option is ignored if \code{cy3pe = NULL}.}

\item{assay}{a numeric index or string specifying the foreground assay in
both the Alexa488 and (optional) Cy3 PBMExperiment objects.
Note that this assay is only used for the initial background subtraction and
Cy3 normalization steps. All subsequent steps will use the default processed
assays from the previous steps.
(default = \code{SummarizedExperiment::assayNames(pe)[1]})}

\item{params}{a named list specifying any non-default parameters that should be used 
for preprocessing the data. Only entries in the list exactly matching the name
of a preprocessing step are considered valid (\code{"backgroundSubtract"},
\code{"cy3Normalize"}, \code{"spatiallyAdjust"},
\code{"normalizeWithinReplicates"}, or \code{"normalizeAcrossReplicates"}).
For each step, non-default parameters must be specified as a list of named
parameter values. (default = \code{list()}). See below for examples.}

\item{skip}{a string vector specifying any steps to skip during preprocessing. Strings
must exactly match step names as with the \code{params} parameter.
(default = \code{vector("character", 0)})}

\item{verbose}{an integer value specifying the level of verbosity. Must be one of
0, 1, or 2. If 0, no progress messages are printed. If 1, progress messages are
only provided to indicate when each preprocessing step is started and finished.
If 2, in addition to the basic progress messages, \code{verbose = TRUE} is also specified
for all preprocessing functions called internally. (default = \code{1L})}
}
\value{
PBMExperiment containing filtered and normalized Alexa488 probe intensities.
}
\description{
This is the main wrapper function for the core PBM data pre-processing steps
implemented in the upbm package.
}
\examples{
data(hoxc9alexa, package = "upbmData")
data(hoxc9cy3, package = "upbmData")
data(refcy3_8x60k_v1, package = "upbmAux")

\dontrun{
upbmPreprocess(hoxc9alexa, hoxc9cy3, refcy3_8x60k_v1)

upbmPreprocess(hoxc9alexa, NULL, NULL, params = list(backgroundSubtract = FALSE))

upbmPreprocess(hoxc9alexa, hoxc9cy3, refcy3_8x60k_v1,
               params = list(backgroundSubtract = list(assayb = "back")),
               skip = "cy3Normalize")
}

}
\seealso{
\code{\link{backgroundSubtract}}, \code{\link{cy3FitEmpirical}}, \code{\link{cy3FitModel}}, \code{\link{cy3Normalize}}, \code{\link{spatiallyAdjust}}, \code{\link{normalizeWithinReplicates}}, \code{\link{normalizeAcrossReplicates}}
}
\author{
Patrick Kimes
}
