% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxPositionWeights.R
\name{approxPositionWeights}
\alias{approxPositionWeights}
\title{Approximate Positional Bias}
\usage{
approxPositionWeights(se, kmers, nk = 10L, verbose = FALSE, .filter = 1L,
  log_scale = FALSE, offset = 1L, .trim = if (.filter > 0L) {     c(1, 36)
  } else {     NULL })
}
\arguments{
\item{se}{SummarizedExperiment of probe intensities}

\item{kmers}{character vector of k-mers to predict.}

\item{nk}{number of top k-mers to use to estimate positional
trend for each sample. This should be larger for longer
k-mers and smaller for shorter k-mers as the number of probes
per k-mer depends strongly on k. (default = 10L)}

\item{verbose}{logical whether to print extra messages during model fitting
procedure. (default = FALSE)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating weights. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{log_scale}{logical whether to calculate bias on log2 scale
intensities. (default = FALSE)}

\item{offset}{integer offset to add to intensities before performing
log2 scale calculations. (default = 1L)}

\item{.trim}{interger vector of length two specifying start and end
of probe sequence to be used. Default is based on the universal
PBM probe design where only leading 36nt should be used.
Ignored if NULL. (default = c(1, 36))}
}
\value{
A vector of weights of length equal to the M - K + 1, where M is the length
of the trimmed probe sequence, and K is the length of the K-mer sequences.
}
\description{
Probe intensities in protein binding microarrays (PBMs) are
highly dependent not only on the sequence content of each
probe, but also the position of the contained sequences within
the probe. Typically, if a motif is closer to the glass slide,
the binding ability of the TF will be reduced, and the resulting
probe intensity will be lower than if the same motif were
positioned within the probe closer to the free end of the
sequence. This function provides a simple way to estimate
the positional bias of a TF based on the relative probe intensities
for probes containing the top N K-mers, ranked by median K-mer
affinity.
}
\author{
Patrick Kimes
}
