% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay2tidy.R
\name{assay2tidy}
\alias{assay2tidy}
\title{Tidy SummarizedExperiment Object}
\usage{
assay2tidy(se, assay = SummarizedExperiment::assayNames(se)[1],
  long = FALSE, .filter = 1L)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{assay}{string name of the assay to use.
(default = \code{assayNames(se)[1]})}

\item{long}{logical whether to transform data to long format and
include colData in output rather than default wide format with
dimension similar to original SummarizedExperiment object.
(default = FALSE)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to plotting. (default = 1)}
}
\value{
tibble with assay data.
}
\description{
Simple helper function to convert assay format data in SummarizedExperiment
object into a tidy tibble for interactive analysis.
}
\author{
Patrick Kimes
}
