% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPBMTable.R
\name{buildPBMTable}
\alias{buildPBMTable}
\title{Build PBM Sample Table}
\usage{
buildPBMTable(gpr_dir, gpr_type = "Alexa488")
}
\arguments{
\item{gpr_dir}{path to directory containing GPR files.}

\item{gpr_type}{string value specifying type of GPR files to find; must
be one of: "Alexa488", "Cy3", "Masliner", "all". (default = "Alexa488")}
}
\value{
tibble with each row corresponding to a sample GPR file.
}
\description{
Helper function to create table of GPR sample files from specified directory path.
Parsing of file names is based on previously observed patterns and may not be exact.
}
\author{
Patrick Kimes
}
