% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy3FitEmpirical.R
\name{cy3FitEmpirical}
\alias{cy3FitEmpirical}
\title{Compute Cy3 Scaling Factors using Emprical Reference}
\usage{
cy3FitEmpirical(se, refse,
  assay = SummarizedExperiment::assayNames(se)[1], useMean = TRUE,
  standardize = TRUE, threshold = 1L, .filter = 1L)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM Cy3 intensity data.}

\item{refse}{SummarizedExperiment object containing PBM Cy3 reference
intensities.}

\item{assay}{string name of the assay to use. (default = \code{SummarizedExperiment::assayNames(se)[1]})}

\item{useMean}{logical whether to use the probe-level mean, rather than
the probe-level median as the Cy3 reference (default = TRUE)}

\item{standardize}{logical whether to standardize residuals using
MAD (median absolute deviation about the median) intensity
computed for PBM Cy3 reference data. (default = TRUE)}

\item{threshold}{numeric threshold on absolute value of log2 ratio between
observed and expected Cy3 intensities. If \code{standardize} is
specified, the log2 ratios are first scaled before comparing against
the threshold. In this case, the threshold is also scaled by the
median of the probe-level MAD intensities. (default = 1L)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1L)}
}
\value{
original SummarizedExperiment object with additional assays corresponding
to ratio of observed vs. expected probe intensities, and whether probes were
flagged as low-quality based on abs(log2(ratio)) > \code{threshold}.
}
\description{
PBM arrays are scanned twice, once for Cy3-tagged dUTPs to quantify
dsDNA abundance at each probe, and again for the Alexa488-tagged 
protein. The Cy3 scans can be used to first filter out probes which
appear to have poor dsDNA enrichment, and second, to scale Alexa488
intensities to account for differences in dsDNA abundance between
probes.

Since probe intensities in the Cy3 scans are independent of the
loaded Alexa488-tagged protein, Cy3 scans should be comparable across
experiments. Based on this observation, a global Cy3 reference can be
constructed by taking the empirical average over a large number of Cy3
scans. Filtering and scaling can then be performed based on deviations
from this reference. Given a SummarizedExperiment of Cy3 scans and an
empirical reference generated by \code{cy3GenerateRef}, this function
returns the residuals from thereference, and the corresponding
observed-expected ratios for each probe as new assays added to the
original Cy3 SummarizedExperiment object.

The returned SummarizedExperiment object can be passed to \code{cy3Normalize}
to filter low quality probes and/or normalize Alexa488 intensities by the
computed ratios.
}
\seealso{
cy3GenerateRef cy3Normalize
}
\author{
Patrick Kimes
}
