% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy3GenerateRef.R
\name{cy3GenerateRef}
\alias{cy3GenerateRef}
\title{Generate Cy3 empirical reference}
\usage{
cy3GenerateRef(cy3pe, assay = SummarizedExperiment::assayNames(cy3pe)[1],
  offset = 1L, register = TRUE)
}
\arguments{
\item{cy3pe}{a PBMExperiment object containing Cy3 intensity data.}

\item{assay}{a numeric index or string specifying the intensity assay.
(default = \code{SummarizedExperiment::assayNames(cy3pe)[1]})}

\item{offset}{a numeric offset to add to intensities before log2 transforming
to prevent NAs with zero intensities. (default = \code{1L})}

\item{register}{a logical value whether to scale intensities across samples.
(default = TRUE)}
}
\value{
PBMExperiment object with Cy3 probe-level reference metrics.
}
\description{
Given a collection of Cy3 scans, this function estimates an empirical
reference distribution of expected Cy3 intensities per probe. Scans
must all be for the same array design. For each probe, the mean,
median, MAD, and standard deviation of the log2 intensities is computed
across all Cy3 scans.

The estimated reference distributions can be passed to \code{cy3FitEmpirical}
with a PBMExperiment of Cy3 scan intensities to determine probe-level outliers and
scaling factors for the individual Cy3 scans relative to the reference.
}
\details{
By default, samples are first scaled to have a common
median intensity, and the reference probe-level intensities are
computed on the log2 scale.

Scaling is performed such that the median intensity of each
sample is equal to the median sample-median intensity of the original
intensities. This multiplicative scaling is equivalent to an additive
shift on the log2 scale. This behavior can be turned off by specifying
\code{register = FALSE}.

The probe-level summary metrics are calculated on the
log2 scale because Cy3 scans will be
used for filtering and scaling, actions which require examining
fold changes and ratios rather than raw intensity differences.
}
\author{
Patrick Kimes
}
