% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy3Normalize.R
\name{cy3Normalize}
\alias{cy3Normalize}
\title{Perform Cy3 Normalization}
\usage{
cy3Normalize(pe, cy3pe, assay = SummarizedExperiment::assayNames(pe)[1],
  match_by, filter = TRUE, scale = FALSE, verbose = FALSE)
}
\arguments{
\item{pe}{a PBMExperiment object containing Alexa488 intensity data.}

\item{cy3pe}{a PBMExperiment object containing Cy3 deviation values from
using \code{cy3FitEmpirical} or \code{cy3FitModel}.}

\item{assay}{a numeric index or string name specifying the assay to use.
(default = \code{SummarizedExperiment::assayNames(pe)[1]})}

\item{match_by}{a string column name in colData of \code{pe} and \code{cy3pe}
to use for matching scans across the two objects; values of
column must be unique for each scan in each object.}

\item{filter}{logical whether to filter "low quality" flagged probes by setting to NA.
(default = TRUE)}

\item{scale}{logical whether to perform scaling by ratio of observed vs. expected
Cy3 intensity at each probe. (default = FALSE)}

\item{verbose}{a logical value whether to print verbose output during
analysis. (default = FALSE)}
}
\value{
PBMExperiment object.
}
\description{
Uses Cy3 intensities to quantify relative dsDNA material at each probe based
on observed vs. expected intensity ratios.
}
\seealso{
fitCy3Models
}
\author{
Patrick Kimes
}
