% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy3Normalize.R
\name{cy3Normalize}
\alias{cy3Normalize}
\title{Perform Cy3 Normalization}
\usage{
cy3Normalize(se, cy3se, assay_name = "fore", match_by = condition,
  filter = TRUE, scale = FALSE)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR intensity information.}

\item{cy3se}{SummarizedExperiment object containing Cy3 GPR model output from
\code{fitCy3Models}.}

\item{assay_name}{string name of the assay to normalize in \code{se}. (default = "fore")}

\item{match_by}{unquoted name of column in colData of SummarizedExperiments
to use for matching samples across the two experiments; values of
column must be unique for each sample in each experiment.
(default = condition)}

\item{filter}{logical whether to filter "low quality" flagged probes by setting to NA.
(default = TRUE)}

\item{scale}{logical whether to perform scaling by ratio of observed vs. expected
Cy3 intensity at each probe. (default = FALSE)}
}
\value{
SummarizedExperiment object.
}
\description{
Uses Cy3 intensities to quantify relative dsDNA material at each probe based
on observed vs. expected intensity ratios.
}
\author{
Patrick Kimes
}
