% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitProbeDistribution.R
\name{fitProbeDistribution}
\alias{fitProbeDistribution}
\title{Fit Probe-Level Convolution Distributions}
\usage{
fitProbeDistribution(se, assay = SummarizedExperiment::assayNames(se)[1],
  model = c("NormGam", "NormExp"), method = "mle", .filter = 1L)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR intensity information.}

\item{assay}{string name of the assay. (default = \code{SummarizedExperiment::assayNames(se)[1]})}

\item{model}{character string specifying model for fitting intensityies. 
Must be one of "NormGam", "NormExp". (default = "NormGam")}

\item{method}{method used to estimate the parameters for normal+exponential model. See more details from 
\code{limma::normexp.fit}. (default = "mle")}

\item{.filter}{integer specifying level of probe filtering to
perform prior to normalization. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}
}
\value{
A dataframe with rows as condition and cols as parameter
}
\description{
This is a collection of several models to fit a normal+exponential or
normal+gamma convolution model to the probe-level intensities of a PBM experiment.
The normal+exponential convolution model is fit using \code{limma::normexp.fit} and 
the normal+gamma convolution model is fit using \code{NormalGamma::normgam.fit}.
This function returns the model parameter estimates for each sample in a dataframe.
}
\author{
Dongyuan Song, Patrick Kimes
}
