% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerFit.R
\name{kmerFit}
\alias{kmerFit}
\title{Fit k-mer Probe Set Models}
\usage{
kmerFit(se, method = c("dl2", "dl"))
}
\arguments{
\item{se}{SummarizedExperiment of k-mer, probe pairs by \code{kmerAggregate}.}

\item{method}{character name of method to use for estimating cross-probe variance
in each k-mer probe set. Currently, the non-iterative DerSimonian-Laird ("dl")
and two-step Dersimonian-Laird ("dl2") methods are supported. (default = "dl2")}
}
\value{
SummarizedExperiment of estimated k-mer affinities.
}
\description{
After performing probe-level aggregation across samples, this function applies probe
set aggregation to obtain k-mer level estimates of affinity.
}
\author{
Patrick Kimes
}
