% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerEstimate.R
\name{kmerFit}
\alias{kmerFit}
\title{Fit K-mer Probe Set Models}
\usage{
kmerFit(se, conditions = TRUE, contrasts = FALSE, baseline = NULL)
}
\arguments{
\item{se}{K-mer level probe set summaries generated by \code{kmerAggregate} with each
assay corresponding to a different group label.}

\item{conditions}{logical specifying whether K-mer summaries should be returned for
each condition. (default = TRUE)}

\item{contrasts}{logical specifying whether K-mer summaries should be returned for
each condition contrast. (default = FALSE)}

\item{baseline}{string name of baseline condition to compare other conditions against;
if not specified and \code{contrasts = TRUE}, baseline is guessed by looking for
'ref' in any of the conditions. (default = NULL)}
}
\value{
SummarizedExperiment of estimated K-mer affinities and/or differential affinities.
}
\description{
After performing probe-level aggregation across samples, this function applies probe
set aggregation to obtain K-mer level estimates of affinity.
}
\author{
Patrick Kimes
}
