% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerFit.R
\name{kmerFit}
\alias{kmerFit}
\title{Fit k-mer probe set models}
\usage{
kmerFit(pe, kmers = uniqueKmers(8L), positionbias = TRUE,
  method = c("dl2", "dl"), contrasts = TRUE, baseline = NULL,
  outlier_cutoff = qnorm(0.995), outlier_maxp = 0.2, verbose = FALSE)
}
\arguments{
\item{pe}{a PBMExperiment object containing probe-level summarized intensity data
returned by \code{\link{probeFit}}.}

\item{kmers}{a character vector of k-mers. (default = \code{uniqueKmers(8L)})}

\item{positionbias}{a logical value whether to correct for bias due to position
of k-mer along probe sequence. (default = TRUE)}

\item{method}{a character name specifying the method to use for estimating cross-probe
variance in each k-mer probe set. Currently, the two-step DerSimonian-Kacker ("dl2") and
non-iterative DerSimonian-Laird ("dl") methods are supported. (default = "dl2")}

\item{contrasts}{a logical value whether to compute contrasts for all columns against a
specified \code{baseline} column. (default = TRUE)}

\item{baseline}{a character string specifying the baseline condition across \code{pe} columns to
use when calculating contrasts. If not specified and set to NULL, the baseline
value is guessed by looking for ``ref" in the column names of \code{pe}. If a unique
matching value is not found, an error is thrown.
This parameter is ignored when \code{contrasts = FALSE}.
(default = NULL)}

\item{outlier_cutoff}{a numeric threshold used for filtering probes from k-mer
probe sets before fitting each k-mer level model. The threshold is
applied to the absolute value of an approximate robust studentized residual
computed for each probe in each probe set and can be turned off by
setting the value to NULL. By default, approximate 0.5% tails are trimmed.
(default = \code{qnorm(0.995)})}

\item{outlier_maxp}{a numeric threshold on the maximum proportion of probes to filter
for each k-mer probe set according to \code{outlier_cutoff}. This should be
set to a reasonably small value to avoid over-filtering based on the approximate
residual threshold. (default = 0.2)}

\item{verbose}{a logical value whether to print verbose output during analysis. (default = FALSE)}
}
\value{
SummarizedExperiment of estimated k-mer affinities and differences with some or all
of the following assays:

\itemize{
\item \code{"affinityEstimate"}: k-mer affinities.
\item \code{"affinityVariance"}: k-mer affinity variances.
\item \code{"contrastDifference"}: (optional) k-mer differential affinities with \code{baseline} condition.
\item \code{"contrastAverage"}: (optional) k-mer average affinities with \code{baseline} condition.
\item \code{"contrastVariance"}: (optional) k-mer differential affinity variances.
}
}
\description{
Given a PBMExperiment of probe-level summaries returned by \code{probeFit} and a list of k-mers,
this function applies probe set aggregation to obtain k-mer level estimates of affinity and
variance on the scale of log2 signal intensities. Additionally, if \code{contrasts = TRUE},
estimates are also estimated for differential k-mer affinity against a baseline condition. 

The output can be passed to \code{\link{kmerTestContrast}}, \code{\link{kmerTestAffinity}},
\code{\link{kmerTestSpecificity}} to perform various statistical tests with the estimated
k-mer statistics.
}
\details{
By default, probe intensities are corrected within each k-mer probe set
to account for biases introduced by where the k-mer lies along the probe sequence. Bias
correction is performed such that the mean cross-probe intensity for each k-mer is
(typically) unchanged. This bias correction step only serves to reduce the cross-probe variance 
and improve downstream inference for each k-mer.

For many low affinity k-mers, probe sets may include several probes with high intensity due to
the k-mer sharing a probe with a separate high affinity k-mer. These probes do not provide
meaningful affinity information for the lower affinity k-mer. To adjust for this possibility,
outlier probes are filtered from each k-mer probe set prior after position bias correction, but
before aggregation. Probes with large approximate studentized residuals are
filtered from each probe set according to a user-specified threshold (\code{outlier_cutoff}).
However, to prevent overfiltering, a maximum proportion of probes to filter from any probe set
should also be specified (\code{outlier_maxp}).

After bias correction and probe filtering, a meta analysis model is fit to each probe set.
Under this model, cross-probe variances are estimated using either the DerSimonian and Kacker (2007)
or DerSimonian and Laird (1986) estimator. The estimated k-mer affinities and variances are
included in the returned SummarizedExperiment as two assays, \code{"affinityEstimate"} and
\code{"affinityVariance"}.

If \code{contrast = TRUE}, k-mer differential affinities, the corresponding variances, and
average affinities are also returned as three assays, \code{"contrastDifference"},
\code{"contrastVariance"}, and \code{"contrastAverage"}. Positive differential affinities indicate
higher affinity relative to the baseline condition.
}
\references{
If using \code{method = "dl2"} cross-probe variance estimator:
\itemize{
\item DerSimonian, R., & Kacker, R. (2007). Random-effects model for meta-analysis of clinical trials: an update. Contemporary Clinical Trials, 28(2), 105-114.
}
If using \code{method = "dl"} cross-probe variance estimator:
\itemize{
\item DerSimonian, R., & Laird, N. (1986). Meta-analysis in clinical trials. Controlled Clinical Trials, 7(3), 177-188.
}
Cross-probe variance estimation code adapted from:
\itemize{
\item Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1-48. URL: http://www.jstatsoft.org/v36/i03/
}
}
\seealso{
\code{\link{probeFit}}, \code{\link{uniqueKmers}}
}
\author{
Patrick Kimes
}
