% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerTest.R
\name{kmerTest}
\alias{kmerTest}
\title{Test k-mers Differences}
\usage{
kmerTest(se, baseline = NULL)
}
\arguments{
\item{se}{SummarizedExperiment of K-mer affinities.}

\item{baseline}{string name of baseline condition to compare other conditions against.
If NULL, baseline is guessed by looking for 'ref' in any of the conditions.
(default = NULL)}
}
\value{
SummarizedExperiment of estimated k-mer affinity differences.
}
\description{
After estimating k-mer affinities from probe sets using \code{kmerFit},
this function computes affinity differences across conditions, relative to
a baseline condition (typically the reference condition).
}
\author{
Patrick Kimes
}
