% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerTest.R
\name{kmerTest}
\alias{kmerTest}
\title{Aggregate Probe-Level Tests to K-mers}
\usage{
kmerTest(se, assay_name, kmers, .filter = 1L, .trim = if (.filter > 0L)
  {     c(1, 36) } else {     NULL }, ...)
}
\arguments{
\item{se}{probe-level testing results}

\item{kmers}{character vector of k-mers to predict.}

\item{.trim}{interger vector of length two specifying start and end
of probe sequence to be used. Default is based on the universal
PBM probe design where only leading 36nt should be used. 
Ignored if NULL. (default = c(1, 36))}

\item{verbose}{logical whether to print extra messages during model fitting
procedure. (default = FALSE)}
}
\value{
tibble of K-mer results.
}
\description{
After performing probe-level testing, this function applies probe set
aggregation to obtain K-mer level inference.
}
\author{
Patrick Kimes
}
