% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerTestAffinity.R
\name{kmerTestAffinity}
\alias{kmerTestAffinity}
\title{Test for k-mer preferential affinities}
\usage{
kmerTestAffinity(se)
}
\arguments{
\item{se}{a SummarizedExperiment of k-mer-level estimated affinities returned by
\code{\link{kmerFit}}.}
}
\value{
SummarizedExperiment of k-mer preferential affinity results with the following
assays:

\itemize{
\item \code{"affinityEstimate"}: input k-mer affinities.
\item \code{"affinityVariance"}: input k-mer affinity variances.
\item \code{"affinitySignal"}: estimated affinity signals above background.
\item \code{"affinityZ"}: studentized signals (\code{affinitySignal / sqrt(affinityVariance)}).
\item \code{"affinityP"}: one-sided tail p-values for studentized signals.
\item \code{"affinityQ"}: FDR-controlling Benjamini-Hochberg adjusted p-values.
}
}
\description{
Using estimated k-mer level affinities returned by \code{kmerFit}, this
function tests for preferential affinity within conditions for each k-mer.
Here, preferential affinity is defined as statistically significant
affinity above a normal background distribution fit to each
condition.

For each condition, the k-mer affinities above background are determined by
first fitting a normal+exponential convolution to the distribution of estimated
affinities. Then, the expected exponential component for each k-mer is taken
as the estimate of preferential affinity. Under this formulation, the normal component
of the fitted convolution corresponds to the background distribution of binding
affinities for k-mers, while the exponential component corresponds to the signal
(preferential) affinity above this background.
}
\author{
Patrick Kimes
}
