% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerTestAffinity.R
\name{kmerTestAffinity}
\alias{kmerTestAffinity}
\title{Test k-mer Affinities}
\usage{
kmerTestAffinity(se)
}
\arguments{
\item{se}{SummarizedExperiment of k-mer results from \code{kmerFit}.}
}
\value{
SummarizedExperiment of significance results for testing k-mer affinity
strength within conditions.
}
\description{
After estimating k-mer level affinities using probe-set aggregate, this
function tests for high affinity within individual conditions.
}
\author{
Patrick Kimes
}
