% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowertailNormalization.R
\name{lowertailNormalization}
\alias{lowertailNormalization}
\title{Lower-Tail Normalization (TESTING)}
\usage{
lowertailNormalization(se, assay_name = "fore", q = 0.4,
  stratify = condition, baseline = NULL, log_scale = FALSE,
  shift = TRUE, method = c("regression", "pca", "quantreg", "quantile",
  "normal"), regtype = 1L, .fits = FALSE, .filter = 1L)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR intensity information.}

\item{assay_name}{string name of the assay to normalize. (default = "fore")}

\item{q}{percentile between 0 and 1 specifying lower tail to use for
fitting truncated normal distribution. (default = 0.4)}

\item{stratify}{unquoted name of column in colData of SummarizedExperiment (or
'\code{sample}') to use for comparing samples; values in column must be
unique for each sample. (default = condition)}

\item{baseline}{string name of baseline condition to compare other conditions
against; if not specified, guessed by looking for
'ref' in any value of the stratifying variable. (default = NULL)}

\item{log_scale}{logical whether cross-sample normalization should be applied
using log2-scaled intensities rather than raw intensities. (default = FALSE)}

\item{shift}{logical whether to apply linear shift before/after scaling.
(default = TRUE).}

\item{method}{character string specifying method to use for normalization.
Must be one of "regression", "pca", "quantreg", "quantile"  or "normal".
(default = "regression")}

\item{.fits}{logical whether to just return a table of fits rather
than the normalized SummarizedExperiment object. (default = FALSE)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to normalization. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}
}
\value{
SummarizedExperiment object with normalized intensities in
new assay, \code{scaled}.
}
\description{
This is a collection of experimental cross-sample normalization procedures which
all attempt to align samples based on lower-tail of probe intensities (or log2-intensities).
Inherent to all of these cross-sample normalization approaches is the assumption that the
lower-tail probes of all samples are primarily background noise, and should therefore be
similiarly distributed, regardless of specific binding affinity or specificity for each
protein. These methods assume that the primary reason cross-sample normalization is needed is
due to differences in protein concentration, which can be corrected by scale and shift
transformation when concentration differences are small.
}
\details{
Currently, 3 appraoches for scale and shift transformation are supported: "regression",
"pca" and "normal."  The "regression" approach takes the lower \code{q}-quantile probes of the
reference sample and fits an ordinary least squares (OLS) model for each non-reference sample (Y)
on the reference sample (X). The fitted intercept and slope parameters are used as the shift and
scaling parameters. The "pca" approach applies principal components analysis (PCA) to the lower-tail
probes for each pair of the reference and a non-reference sample, and the first PC is used as the
scaling parameter. The shift parameter is calculated such that the first PC passes through the
means of both samples. The "quantreg" method fits a quantile regression model rather than the
OLS model used for the "regression" method. The "quantile" method simply scales distributions to
have the same value at the specified quantile. This reduces to the "scale" method in the
\code{limma::normalizeBetweenArrays} function. For the "normal" method, a background (normal) plus
signal is assumed for probe-level intensities, and the normal background mean is estimated using
the \code{bg.parameters} function from the \code{affy} package. The corresponding
standard deviation is estimated separately from the lower tail for each sample by
fitting a truncated normal distribution to the tail.

\emph{This code is still under testing and not a finalized normalization procedure}.
}
\author{
Dongyuan Song, Patrick Kimes
}
