% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapkmers.R
\name{mapkmers}
\alias{mapkmers}
\title{Map K-mer Motifs to PBM Probe Sequences}
\usage{
mapkmers(probes, kmers)
}
\arguments{
\item{probes}{table containing sequence information for probes,
e.g. read in from "8x60k_v1_sequences.txt". At a minimum,
must have a column called "Sequence" with the array probe
sequences.}

\item{kmers}{character vector of K-mers (of equal length) to which
probes should be mapped.}
}
\value{
a table with 8mer sequences (seq), and probes, where each row
corresponds to a unique occurrence of the sequence on a probe and on
the array.
}
\description{
This helper function takes a standardized PBM sequence design file,
e.g. "8x60k_v1_sequences.txt", and k-mer sequences of interest,
and identifies the subset of probes containing each k-mer sequence.
The output table can be used to compute k-mer-level affinities from
the probe-level intensity values.
}
\author{
Patrick Kimes
}
