% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM-accessors.R
\docType{methods}
\name{PBMclass-accessors}
\alias{PBMclass-accessors}
\alias{probeFilter}
\alias{probeTrim}
\alias{probeCols}
\alias{probeFilter<-}
\alias{probeTrim<-}
\alias{design<-,PBMDesign,data.frame-method}
\alias{design,PBMDesign-method}
\alias{SummarizedExperiment,PBMExperiment-method}
\title{PBM slot accessors and setters}
\usage{
probeFilter(object)

probeTrim(object)

probeCols(object)

probeFilter(object) <- value

probeTrim(object) <- value

\S4method{design}{PBMDesign,data.frame}(object, ...) <- value

\S4method{design}{PBMDesign}(object)

\S4method{SummarizedExperiment}{PBMExperiment}(assays)
}
\arguments{
\item{object}{a PBMDDesign or PBMExperiment object.}

\item{value}{a new value to be assigned to the corresponding components
of object.}

\item{...}{further arguments, perhaps used by methods.}

\item{assays}{a PBMExperiment object. Only used for converting
PBMExperiment to SummarizedExperiment object.}
}
\value{
slot value for accessors, modified object for setters.
}
\description{
Simple functions for accessing and setting specialized slots
for PBM associated classes (PBMDesign, PBMExperiment).
}
\author{
Patrick Kimes
}
