% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMExperiment-class.R
\docType{class}
\name{PBMExperiment-class}
\alias{PBMExperiment-class}
\title{PBMExperiment class}
\description{
This class is an extension of the \code{\link{SummarizedExperiment}} class to
store protein binding microarray data.
}
\author{
Patrick Kimes
}
