% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMExperiment-constructor.R
\name{PBMExperiment}
\alias{PBMExperiment}
\title{Create a new PBMExperiment object}
\usage{
PBMExperiment(x, useMean = FALSE, filterFlags = TRUE,
  readBackground = TRUE, probes = NULL)
}
\arguments{
\item{x}{table of scans with at least a single `gpr' column
corresponding to GPR file paths.}

\item{useMean}{logical whether to use mean fluorescent intensity
for each probe rather than median fluorescent intensity.
(default = FALSE)}

\item{filterFlags}{logical whether to replace intensity values at probes
flagged manually or automatically as being low quality
 (`Bad': -100, `Absent': -75, `Not Found': -50) with NA. (default = TRUE)}

\item{readBackground}{logical whether to also read in probe background
intensities. (default = TRUE)}

\item{probes}{data.frame containing probe sequences in a column, `Sequence',
or a character vector specifying the probe sequences. If specified,
these values will be added to the rowData of the returned
SummarizedExperiment object. (default = NULL)}
}
\value{
\code{\link[PBMExperiment-class]{PBMExperiment}} object with assays containing
intensities for the scans specified in the table of scans. The object
includes the following assays:
* \code{fore} - foreground probe intensities,
* \code{back} - background probe intensities (unless \code{readBackground = FALSE}).

Scan metadata included in the table of scans is stored as column data.
Probe design information, if specified, is included as row data.
}
\description{
Read PBM data from a table containing paths to GPR files.
}
\details{
If raw data files (cleaned GPR data files, e.g. uploaded to UniPROBE) include
probe ID, Name, or Sequence information, this is compared across samples and
an error is thrown if the probe-to-sequence mapping is not common across all
samples.
}
\seealso{
\code{\link{PBMExperiment-class}}
}
\author{
Patrick Kimes
}
