% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMExperiment-constructor.R
\name{PBMExperiment}
\alias{PBMExperiment}
\title{Create a new PBMExperiment object}
\usage{
PBMExperiment(..., pbmDesign = NULL, probeFilter = list(),
  probeTrim = numeric(), probeCols = c("Sequence", "probeID"))
}
\arguments{
\item{...}{an optional SummarizedExperiment containing PBM probe intensity data and
corresponding annotations or parameters to be passed to the SummarizedExperiment
constructor.}

\item{pbmDesign}{an optional PBMDesign containing design information. If specified,
\code{probeFilter}, \code{probeTrim}, and \code{probeCols} parameters will be 
ignored. Ignored if NULL. (default = NULL)}

\item{probeFilter}{an optional named list of probe filters to be used to subset
probes during data analysis steps. List names must correspond to columns in
rowData. List entries must be single-parameter functions to be called on the
corresponding column to return a logical vector of probes to keep (TRUE) or
drop (FALSE) during analysis. (default = \code{list()})}

\item{probeTrim}{an optional integer vector of length 2 specifying start and end
positions in probe `Sequence' to use in analysis steps. (default = \code{numeric()})}

\item{probeCols}{an optional character vector of rowData column names corresponding
to probe design information. (default = \code{c("Sequence", "probeID")})}
}
\value{
\code{\link[=PBMExperiment-class]{PBMExperiment}} object.
}
\description{
Create a PBMExperiment object by directly specifying optional probe annotation
information (\code{probeFilter}, \code{probeTrim}, \code{probeCols}), along with
a based SummarizedExperiment object containing probe intensity data.
}
\seealso{
\code{\link{PBMExperiment-class}}, \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment-class}}, \code{\link{gpr2PBMExperiment}}
}
\author{
Patrick Kimes
}
