% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMExperiment-constructor.R
\name{PBMExperiment}
\alias{PBMExperiment}
\title{Load raw PBM scan data as PBMExperiment}
\usage{
PBMExperiment(x, useMedian = TRUE, filterFlags = TRUE,
  readBackground = TRUE, probes = NULL, ...)
}
\arguments{
\item{x}{a data.frame of scans with at least a single `gpr' column
corresponding to GPR file paths.}

\item{useMedian}{a logical value whether to use median fluorescent intensity
for each probe rather than mean fluorescent intensity.
(default = TRUE)}

\item{filterFlags}{a logical value whether to replace intensity values at probes
flagged manually or automatically as being low quality
 (`Bad': -100, `Absent': -75, `Not Found': -50) with NA. (default = TRUE)}

\item{readBackground}{a logical value whether to also read in probe background
intensities. (default = TRUE)}

\item{probes}{a \code{\link[PBMDesign-class]{PBMDesign}} object containing
the probe design used for the data. (default = NULL)}

\item{...}{optionally, probe design information can be directly specified using
\code{design=}, \code{probeFilter=}, \code{probeTrim=} parameters. These
will be ignored if \code{probes=} is specified.}
}
\value{
\code{\link[PBMExperiment-class]{PBMExperiment}} object with assays containing
intensities from the GPR files specified in the table of scans. The object
includes the following assays:
* \code{fore} - foreground probe intensities,
* \code{back} - background probe intensities (unless \code{readBackground = FALSE}).

Scan metadata included in the data.frame are stored as column data.
Probe design information, if specified, is included as row data.
}
\description{
Read PBM scan data from GenePix Results (GPR) files and combine into
a single \code{\link[PBMExperiment-class]{PBMExperiment}} object.
Paths to GPR file must be specified as part of a larger
}
\details{
GPR files are parsed differently depending on the file type which can take values
of ``Alexa", ``Cy3", or ``Masliner". Both ``Alexa" and ``Cy3" scans
are treated as untouched GenePix Results (GPR) files, while ``Masliner" files are
treated as Masliner-processed GPR files.

If a `scan' column is not provided in the input table, GPR files are assumed to
be Alexa488 scans unless the file name in the `gpr' column contains "Masliner".

If ``raw data" files (processed GPR data files, e.g. uploaded to UniPROBE)
include probe ID, Name, or Sequence information, this is compared across samples and
an error is thrown if the probe-to-sequence mapping is not common across all
samples.
}
\seealso{
\code{\link{PBMExperiment-class}}
}
\author{
Patrick Kimes
}
