% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmFilterProbes.R
\name{pbmFilterProbes}
\alias{pbmFilterProbes}
\title{Filter PBM Probes}
\usage{
pbmFilterProbes(se, assay_name, level = 0L)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR
intensity information.}

\item{assay_name}{string name of the assay to plot.
(default = "gpr")}

\item{level}{integer specifying level of probe filtering to
perform prior to plotting. (default = 0)}
}
\value{
SummarizedExperiment object
}
\description{
Not all probes in a PBM experiment should be used for analysis.
Given a SummarizedExperiment containing PBM intensities,
this functions provides multiple levels of filtering to remove
uninformative and low quality probes based on quality control
metrics.
}
\details{
The function supports the following levels of filtering.
All filtering is progressive, i.e. setting level = 2 will
also include filtering at level = 1.
\itemize{
\item{0}{no filtering}
\item{1}{filter on de Bruijn probes}
}
}
\author{
Patrick Kimes
}
