% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmPlotChip.R
\name{pbmPlotChip}
\alias{pbmPlotChip}
\title{Plot of Intensities on Grid}
\usage{
pbmPlotChip(se, assay_name = "fore", log_scale = TRUE, .facet = TRUE,
  .filter = 1)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR
intensity information.}

\item{assay_name}{string name of the assay to plot.
(default = "fore")}

\item{log_scale}{logical whether to plot the intensities
on the log-scale. (default = TRUE)}

\item{.facet}{logical whether plot should be faceted using
the default 'condition' column from the colData of the
SummarizedExperiment. (default = TRUE)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to plotting. (default = 1)}
}
\value{
ggplot object
}
\description{
Probe-level intensities plotted by row and column indices
defined in the \code{rowData} of the SummarizedExperiment
object. This can be useful for visually inspecting spatial
trends and biases across the scanned chip.
}
\details{
If the values in the specified assay include negative
values, \code{log_scale = FALSE} will be used. Note that
this behavior is different from the other plotting functions
in this package. To get around this, manually set all negative
values to NA prior to plotting.
}
\author{
Patrick Kimes
}
