% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictKmers.R
\name{predictKmers}
\alias{predictKmers}
\title{Estimate K-mer Affinities}
\usage{
predictKmers(se, assay_name = "fore", kmers = NULL, offset = 1,
  verbose = FALSE, .filter = 1L, .trim = if (.filter > 0L) {     c(1,
  36) } else {     NULL }, ...)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data.}

\item{assay_name}{string name of the assay to use. (default = "fore")}

\item{kmers}{character vector of k-mers to predict.}

\item{offset}{integer offset to add to intensities before log2 scaling to
prevent errors with zero intensities. If set to 0, probes with
zero intensities are dropped/ignored in estimation. (default = 1)}

\item{verbose}{logical whether to print extra messages during model fitting
procedure. (default = FALSE)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{.trim}{interger vector of length two specifying start and end
of probe sequence to be used. Default is based on the universal
PBM probe design where only leading 36nt should be used.
Ignored if NULL. (default = c(1, 36))}

\item{...}{parameters to be passed to glmnet call - will overwrite
default paramaters.}
}
\value{
SummarizedExperiment object with predicted affinities.
}
\description{
This function uses elastic net regularized least squares regression
to estimate k-mer affinities from PBM data. The base model
estimates affinities across multiple samples with a shared probe-level
effect. Estimates are reported on the scale of log2 relative affinity.
}
\details{
The default signature for the glmnet call is:
\itemize{
\item family = 'gaussian'
\item alpha = 0.5
\item intercept = FALSE
\item lambda = rev(0:50)
\item lower.limits = 0
\item thresh = 1e-10
}

Features to be added:
\itemize{
\item option to select different regression models
\item option to return glmnet results for all alpha
\item frozen RMA-like pooled estimate?
}
}
\author{
Patrick Kimes
}
