% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictKmers.R
\name{predictKmers}
\alias{predictKmers}
\title{Estimation of K-Mer Intensities}
\usage{
predictKmers(se, kmers = NULL, assay_name = "fore", std_array = TRUE,
  offset = 1, verbose = FALSE, ...)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data.}

\item{kmers}{character vector of k-mers to predict.}

\item{assay_name}{string name of the assay to use. (default = "fore")}

\item{std_array}{logical whether the array is a standard PBM array.
See Details for more information. (default = TRUE)
(default = TRUE)}

\item{offset}{integer offset to add to intensities before log2 scaling to
prevent errors with zero intensities. (default = 1)}

\item{verbose}{logical whether to print extra messages during model fitting
procedure. (default = FALSE)}

\item{...}{parameters to be passed to glmnet call - will overwrite
default paramaters.}
}
\value{
SummarizedExperiment object with predicted affinities.
}
\description{
This function uses stuff to do stuff.
}
\details{
If \code{stdAssay} is TRUE, prior to estimating k-mer affinities,
probes are subset to de Bruijn sequences as determined by the ID column
of the rowData, and probe sequences are trimmed to only use the unique
36nt region.

The default signature for the glmnet call is:
\itemize{
\item family = 'gaussian'
\item alpha = 0.5
\item intercept = FALSE
\item lambda = rev(0:50)
\item lower.limits = 0
\item thresh = 1e-10
}

Features to be added:
\itemize{
\item option to select different regression models
\item option to return glmnet results for all alpha
\item frozen RMA-like pooled estimate?
}
}
\author{
Patrick Kimes
}
