% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerTest.R
\name{probeTest}
\alias{probeTest}
\title{Test Probe Intensity Differences}
\usage{
probeTest(se, design, assay_name = "fore", offset = 1L, .filter = 1L,
  baselines = NULL)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data.}

\item{design}{formula specifying design of test, where variables are
taken from the \code{colData} of \code{se}.}

\item{assay_name}{string name of the assay to use. (default = "fore")}

\item{offset}{integer offset to add to intensities before log2 scaling to
prevent errors with zero intensities. If set to 0, probes with
zero intensities are dropped/ignored in estimation. (default = 1)}

\item{.filter}{integer specifying level of probe filtering to
perform prior to estimating affinities. See \code{pbmFilterProbes}
for more details on probe filter levels. (default = 1)}

\item{baselines}{named list of baseline conditions for any categorical
variables included in the model. Ignored if NULL. (default = NULL)}
}
\value{
SummarizedExperiment object with probe-level testing results. Each assay
contains the results for a single coefficient in the model specified
in \code{design}.
}
\description{
This function uses limma to perform probe-level tests of intensity differences
across conditions. Contrasts should be specified as a design formula.
}
\examples{
\dontrun{
res <- probeTest(mygpr, ~ 1 + condition)
}

}
\author{
Patrick Kimes
}
