% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPBMExperiment.R
\name{readGPR}
\alias{readGPR}
\title{Read GPR File as Assay}
\usage{
readGPR(gpr_path, gpr_type, useMean = FALSE, filterFlags = TRUE,
  readBackground = TRUE)
}
\arguments{
\item{gpr_path}{path to GPR file.}

\item{gpr_type}{scan type; one of "Alexa488", "Cy3", "Masliner".}

\item{useMean}{logical whether to use mean fluorescent intensity
for each probe rather than median fluorescent intensity.
(default = FALSE)}

\item{filterFlags}{logical whether to replace intensity values at probes
flagged manually or automatically as being low quality.
 ('Bad': -100, 'Absent': -75, 'Not Found': -50) with NA. (default = TRUE)}

\item{readBackground}{logical whether to also read in probe background
intensities. (default = TRUE)}
}
\value{
tibble (data.frame-like) object of a single GPR file with three or four
columns: 'Column', 'Row', 'foreground', (optionally) 'background'.
('ID' and 'Name' columns are ignored as these may be incorrect in the GPR file.)
}
\description{
Helper function for reading in a single GPR file.
}
\details{
Since the names of foreground and background columns can vary across samples,
they are identified using regular expressions. Therefore, if the column names
deviate from the expected format ("F.* Mean", "F.* Median". "B.* Mean", "B.* Median"),
the function may fail to read the foreground and background intensities.
}
\author{
Patrick Kimes
}
