% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPBMExperiment.R
\name{readGPR}
\alias{readGPR}
\title{Read GPR File as Assay}
\usage{
readGPR(x, useMean = FALSE, useBackground = FALSE, filterFlags = TRUE)
}
\arguments{
\item{x}{path to GPR fle}

\item{useMean}{logical whether to use mean fluorescent intensity
for each probe rather than median fluorescent intensity
(default = FALSE)}

\item{useBackground}{logical whether to use background subtracted
intensity rather than non-subtracted intensity
(default = FALSE)}

\item{filterFlags}{logical whether to replace intensity values at probes
flagged manually or automatically as being low quality
 ('Bad': -100, 'Absent': -75, 'Not Found': -50) with NA. (default = TRUE)}
}
\value{
tibble (data.frame-like) object of a single GPR file with three
columns: 'Column', 'Row', 'intensity'. ('ID' and 'Name' columns are
ignored as these may be incorrect in the GPR file.)
}
\description{
Helper function for reading in a single GPR file.
}
\details{
Since the name of the value column can vary across samples,
we pull it out using its index. Therefore, if the order or number of
columns in the GPR file is altered, reading may fail.
}
\author{
Patrick Kimes
}
