% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiallyAdjust.R
\name{spatiallyAdjust}
\alias{spatiallyAdjust}
\title{Spatial Adjustment of Samples}
\usage{
spatiallyAdjust(pe, assay = SummarizedExperiment::assayNames(pe)[1],
  k = 15, returnBias = TRUE, verbose = FALSE)
}
\arguments{
\item{pe}{a PBMExperiment object containing PBM intensity data.}

\item{assay}{a string name of the assay to adjust. (default = \code{SummarizedExperiment::assayNames(pe)[1]})}

\item{k}{an odd integer specifying size of region to use to for computing
local bias. (default = 15)}

\item{returnBias}{a logical whether to include the spatial bias as an
additional 'assay' (called 'spatialbias') in the returned
SummarizedExperiment object. (default = TRUE)}

\item{verbose}{a logical value whether to print verbose output during
analysis. (default = FALSE)}
}
\value{
Original PBMExperiment object with additional assay corresponding to
spatially adjusted intensities and a new column added to the colData,
`spatialMedian', containing the global median intensity of the original
probe-level data used to compute spatial bias for each sample.
}
\description{
Given PBM intensities stored as a SummarizedExperiment, this function
computes the local intensity bias for each probe, and returns the
same SummarizedExperiment object with each probe corrected (scaled) by the bias.
Optionally, the per-probe bias can also be returned as an additional array
in the bias corrected SummarizedExperiment object. The spatial bias at each
probe is defined as the ratio between the median intensity in a `k` by `k`
square region surrounding the probe, and the median intensity of all probes
across the array. This approach is taken directly from the original PBM
analysis pipeline described in Berger and Bulyk (Nature Protocols, 2008).
The size of the local region can be specified by the user, with a default
size of 15 x 15 (as used in the aforementioned publication).
}
\author{
Patrick Kimes
}
