% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiallyAdjust.R
\name{spatiallyAdjust}
\alias{spatiallyAdjust}
\title{Spatial Adjustment of Samples}
\usage{
spatiallyAdjust(se, assay_name = "fore", k = 15, returnBias = TRUE,
  log_scale = FALSE, .force = FALSE, .nonnegative = TRUE)
}
\arguments{
\item{se}{SummarizedExperiment object containing PBM intensity data}

\item{assay_name}{string name of the assay to adjust. (default = "fore")}

\item{k}{odd integer specifying size of region to use to for computing
local bias (default = 15)}

\item{returnBias}{logical whether to include the spatial bias as an
additional 'assay' (called 'spatialbias') in the returned
SummarizedExperiment object.}

\item{log_scale}{logical whether to perform adjustment on log2 scale
intensities. Returned values will be transformed back to non-log scale.
STILL TESTING. (default = FALSE)}

\item{.force}{logical whether to run adjustment even if data
has already been spatially adjusted. (default = FALSE)}

\item{.nonnegative}{logical whether to restrict intensities to non-negative
values by setting negative values to NA. (default = TRUE)}
}
\value{
SummarizedExperiment object with spatially adjusted intensities.
Row and column metadata are copied from the original SummarizedExperiment
object.
}
\description{
Given PBM intensities stored as a SummarizedExperiment, this function
computes the local intensity bias for each probe, and returns the
same SummarizedExperiment object with the bias subtracted from each probe.
Optionally, the per-probe bias can also be returned as an additional array
in the bias corrected SummarizedExperiment object. The spatial bias at each
probe is defined as the difference between the median intensity in a `k` by `k`
square region surrounding the probe, and the median intensity of all probes
across the array. This approach is taken directly from the original PBM
analysis pipeline described in Berger and Bulyk (Nature Protocols, 2008).
The size of the local region can be specified by the user, with a default
size of 15 x 15 (as used in the aforementioned publication).
}
\author{
Patrick Kimes
}
