% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeKmers.R
\name{summarizeKmers}
\alias{summarizeKmers}
\title{Summarize K-mer Intensities}
\usage{
summarizeKmers(se, assay_name = "fore", kmers = NULL, verbose = FALSE,
  .filter = 1L, .trim = if (.filter > 0L) {     c(1, 36) } else {     NULL
  }, ...)
}
\arguments{
\item{se}{SummarizedExperiment object containing GPR
intensity information.}

\item{assay_name}{string name of the assay to plot.
(default = "fore")}

\item{kmers}{character vector of k-mers to predict.}

\item{assay_name}{string name of the assay to use. (default = "fore")}

\item{std_array}{logical whether the array is a standard PBM array.
See Details for more information. (default = TRUE)
(default = TRUE)}
}
\value{
SummarizedExperiment object with summarized intensity information in
separate assays.
}
\description{
This function calculates k-mer intensity information from
PBM data.
}
\author{
Patrick Kimes
}
