% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeKmers.R
\name{summarizeKmers}
\alias{summarizeKmers}
\title{Compute K-mer summary metrics from probe intensities}
\usage{
summarizeKmers(
  pe,
  assay = SummarizedExperiment::assayNames(pe)[1],
  kmers = uniqueKmers(8L),
  metrics = c("median", "mean", "mad", "sd", "log2mean", "log2mad", "log2sd", "na",
    "quantile"),
  offset = 1,
  q = 0.25
)
}
\arguments{
\item{pe}{a PBMExperiment object containing PBM intensity data.}

\item{assay}{a string name of the assay to adjust. (default = \code{SummarizedExperiment::assayNames(pe)[1]})}

\item{kmers}{a character vector of k-mers to predict. (default = \code{uniqueKmers(8L)})}

\item{metrics}{a character vector of statistics to calculate for each scan.
The set of supported statistics are listed in the details. By default,
all possible statistics are computed, but this is not recommended as
it can be computational expensive. (default = \code{c("median", "mean", "mad", "sd", "log2mean", "log2mad", "log2sd", "na", "quantile")})}

\item{offset}{an integer offset to add to intensities before log2 scaling to
prevent errors with zero intensities. If set to 0, probes with
zero intensities are dropped/ignored for log-scaled metrics. (default = 1)}

\item{q}{a vector of quantiles which should be used if \code{"quantile"} is specified as part
of \code{metrics}. (default = 0.25)}
}
\value{
SummarizedExperiment object with intensity information summarized for
k-mers. Each metric is returned as a separate assay, with each rows corresponding
to k-mers. k-mer sequences and the number of corresponding probes are given
in the rowdata of the object.
}
\description{
This function calculates k-mer intensity information from
PBM data.
}
\details{
The following statistics are currently supported.
* \code{"median"}: median probe intensity
* \code{"mean"}: mean probe intensity
* \code{"mad"}: MAD probe intensity (scaled by 1.4826)
* \code{"sd"}: SD probe intensity
* \code{"log2mean"}: mean probe log2(intensity + offset)
* \code{"log2mad"}: MAD probe log2(intensity + offset)
* \code{"log2sd"}: SD probe log(intensity + offset)
* \code{"na"}: number of NA probes
* \code{"quantile"}: q-quantile probe intensity
}
\author{
Patrick Kimes
}
