% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-PBMExperiment.R
\name{tidy-SummarizedExperiment}
\alias{tidy-SummarizedExperiment}
\alias{tidy.SummarizedExperiment}
\title{Tidy SummarizedExperiment object}
\usage{
\method{tidy}{SummarizedExperiment}(x, assay = SummarizedExperiment::assayNames(x)[1], long = FALSE, ...)
}
\arguments{
\item{x}{a SummarizedExperiment object.}

\item{assay}{a numeric index or string name specifying the assay to tidy.
If multiple assays are specified, assays will be combined
as separate columns.
(default = \code{SummarizedExperiment::assayNames(x)[1]})}

\item{long}{a logical whether to transform data to long format and
include colData in output rather than default wide format with
dimension similar to original PBMExperiment object.
Ignored and set to TRUE if \code{assay} specifies more than a
single assay. (default = FALSE)}

\item{...}{other parameters for the \code{tidy} generic function.}
}
\value{
tibble containing values from a single SummarizedExperiment assay
along with rowData and optionally colData.
}
\description{
Simple helper function to convert a single assay data in
a SummarizedExperiment object into a tidy tibble. If \code{long = FALSE},
a tibble with the same number of rows as the input SummarizedExperiment
is constructed with columns corresponding to the columns of the
SummarizedExperiment and any rowData. If \code{long = TRUE}, the assay
data is "tidied" such that each row of the tibble corresponds to a single
value of the assay. When \code{long = TRUE}, in addition to rowData, any
colData is also added as additional columns.
}
\author{
Patrick Kimes
}
