% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueKmers.R
\name{uniqueKmers}
\alias{uniqueKmers}
\title{Generate Unique K-mer Sequences}
\usage{
uniqueKmers(k = 8L, .largek = FALSE)
}
\arguments{
\item{k}{integer length of oligonucleotide sequences to be
returned. (default = 8)}

\item{.largek}{logical whether to allow specifying a large k,
i.e. k > 10L. (default = FALSE)}
}
\value{
vector of unique k-mer strings.
}
\description{
This simple function returns a vector of all unique k-mers of
a specified length, filtered to remove reverse complement redundancy
such that, e.g. only one of "AA" and "TT" is kept. When deciding
between a k-mer and the reverse complement, the "smaller" sequence
is kept, where simple alphabetical ordering is used. In the example
above, "AA" will be kept because "AA" <= "TT". For \code{k = 8}, 
this will return the same set of 8-mers as is used in the Universal
PBM Analysis Suite software.
}
\details{
Note, this function is not meant to be used for k > 10, and will return
an error if k > 10 is specified unless, \code{largek = TRUE}.
}
\author{
Patrick Kimes
}
